WIP Twitter bot for timeline events
