+++ src/Contributing.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { HashLink as Link } from "react-router-hash-link";
 import Disqus from 'disqus-react';
 import BackToTop from "./shared/BackToTop";
-import TimelineData from './data/timelineData.js';
+import TimelineData from './data/timelineDataLoader.js';
 import SourceUtils from './util/sourceUtils.js';
 
 export default class Contributing extends React.Component<any, any> {
