--- src/data/timelineData.js
@@ -1,2846 +1,0 @@
-import React from 'react';
-
-export default {
-
-    data: [
-        {
-            description: <span>While attending a Stone Roses gig at International 2 in Manchester, UK, Noel Gallagher meets Graham Lambert of the Inspiral Carpets.</span>,
-            date: 'May 30',
-            source: { name: 'exclaim!*@#', title: 'Oasis (Here\'s The Story) Morning Glory', url: 'http://exclaim.ca/music/article/oasis-_heres_the_story_morning_glory' },
-            title: 'Noel Gallagher Meets Graham Lambert',
-            type: 'noteworthy',
-            year: '1988'
-        },
-        {
-            description: <span>Noel Gallagher auditions to be the signer of the Inspiral Carpets.  While Noel does not get the job, he impresses the band.</span>,
-            date: 'December 21',
-            source: { name: 'Oasis FAQ', title: 'CHRONOLOGY 1988 to 1992', url: 'http://www.oocities.org/sunsetstrip/underground/3284/chron92.html' },
-            title: 'Noel Gallagher Auditions for the Inspiral Carpets',
-            type: 'noteworthy',
-            year: '1988'
-        },
-        {
-            description: <span>Noel Gallagher joins the Inspiral Carpets' road crew as a technician and roadie.  It is in this job that Noel would meet future Oasis collaborator Mark Coyle, who did the Inspirals' live sound.</span>,
-            date: 'May',
-            source: { name: 'Oasis FAQ', title: 'CHRONOLOGY 1988 to 1992', url: 'http://www.oocities.org/sunsetstrip/underground/3284/chron92.html' },
-            title: 'Noel Gallagher Hired by the Inspiral Carpets',
-            type: 'noteworthy',
-            year: '1989'
-        },
-        {
-            description: <span>A precursor to Oasis, Paul "Bonehead" Arthurs and Paul "Guigsy" McGuigan form a band called The Rain.  The band would go on to play a few gigs before disbanding and replacing signer Chris Hutton with Liam Gallagher.</span>,
-            date: 'Early',
-            source: { name: 'Oasis FAQ', title: 'CHRONOLOGY 1988 to 1992', url: 'http://www.oocities.org/sunsetstrip/underground/3284/chron92.html' },
-            title: 'Bonehead and Guigsy Form The Rain',
-            type: 'noteworthy',
-            year: '1991'
-        },
-        {
-            description: <span>First gig as Oasis at The Boardwalk, Manchester, UK.  At this point, the band is a 4-piece made up of Liam Gallagher, Paul "Bonehead" Arthurs, Paul "Guigsy" McGuigan and Tony McCarroll.  The Inspiral Carpets are in attendance, accompanied by roadie Noel Gallagher, who sees his brother's band perform live for the first time.</span>,
-            date: 'August 18',
-            source: { name: 'Manchester Evening News', title: 'Oasis Come Home', url: 'https://www.manchestereveningnews.co.uk/whats-on/music/oasis-come-home-1191304' },
-            title: 'First gig as Oasis',
-            type: 'gigs',
-            year: '1991'
-        },
-        {
-            description: <span>First studio recording done by Oasis at Out of the Blue Studios, Manchester, UK.  While Noel Gallagher is listed as "rythm" on the insert card of the resulting demo tape, he was not present at the session according to Tony McCarroll.</span>,
-            date: 'Autumn',
-            source: { name: 'Oasis Demo Info', title: 'Pre-Definitely Maybe', url: 'https://monobrowdemos.wordpress.com/pre-definitely-maybe/' },
-            title: 'First Recording Session',
-            type: 'recordings',
-            year: '1991'
-        },
-        {
-            description: <span>First gig with Noel Gallagher at the Boardwalk in Manchester, UK.  The band performs four songs including a "cover of a house tune", most likely <i>Better Let You Know</i> which is based on <i>Feel the Groove</i> by Cartouche.</span>,
-            date: 'October 19',
-            source: { name: 'Oasis Interviews', title: 'Noel Gallagher Interview - Q Magazine July 1995', url: 'http://oasisinterviews.blogspot.com/1995/07/noel-gallagher-q-july-1995.html' },
-            title: 'First gig with Noel Gallagher',
-            type: 'gigs',
-            year: '1991'
-        },
-        {
-            description: <span>Uptown Magazine publishes the first article about Oasis by Steve Cowell.  It is also the band's first interview and photo shoot.  The article mentions the song <i>Take Me</i>, which is part of the demos recorded at Out of the Blue studios, and that both Virgin Records and Rough Trade Records have interest in signing the band.</span>,
-            date: 'October 22',
-            source: { name: 'Oasis - Uptown Magazine', title: 'Oasis by Steve Cowell', url: 'http://live4ever.proboards.com/post/1557454' },
-            title: 'First Magazine Article - Uptown Magazine',
-            type: 'noteworthy',
-            year: '1991'
-        },
-        {
-            description: <span>Oasis record a performance for a television event called <i>The Blackpool Roadshow</i>, playing the song <i>Take Me</i>.  Though the performance is never televised it is considered the first Oasis recording for TV.</span>,
-            date: 'July 18',
-            source: { name: 'Oasis Recording Info', title: 'Oasis on TV', url: 'http://www.oasis-recordinginfo.co.uk/?page_id=455' },
-            title: 'Oasis Perform for First Television Appearance',
-            type: 'noteworthy',
-            year: '1992'
-        },
-        {
-            description: <span>Oasis takes part in a recording session at Mark Coyle's home studio at 388 Mauldeth Road West in Manchester, UK over various dates in Spring, 1993.  Songs from this session would appear on later releases, such as <i>Definitely Maybe</i> and anniversary releases of <i>(What's the Story) Morning Glory?</i> and <i>Be Here Now</i>.</span>,
-            date: 'Spring',
-            source: { name: 'Oasis Demo Info', title: 'Definitely Maybe', url: 'https://monobrowdemos.wordpress.com/definitely-maybe-2/' },
-            title: 'Recording Session, Mark Coyle Home Studio, Manchester, UK',
-            type: 'recordings',
-            year: '1993'
-        },
-        {
-            description: <span>Recording session at the Porter Street Studio in Liverpool, UK that produced the "<a href="https://en.wikipedia.org/wiki/Live_Demonstration" target="_blank" rel="noopener noreferrer">Live Demonstration"</a> tape.</span>,
-            date: 'March',
-            source: { name: 'Oasis Demo Info', title: 'Definitely Maybe', url: 'https://monobrowdemos.wordpress.com/definitely-maybe-2/' },
-            title: 'Real People Recording Session, Liverpool, UK',
-            type: 'recordings',
-            year: '1993'
-        },
-        {
-            description: <span>Noel Gallagher meets Johnny Marr after befriending Marr's brother Ian.  Marr takes Noel and Oasis under his wing by helping the band record demos and introduces the band to his manager, Marcus Russell.</span>,
-            date: 'May',
-            source: { name: 'Dean Street Press', title: 'Getting High by Paolo Hewitt', url: 'http://deanstreetpress.co.uk/books/hewi02' },
-            title: 'Noel Gallagher Meets Johnny Marr',
-            type: 'noteworthy',
-            year: '1993'
-        },
-        {
-            description: <span>Oasis forces their way on to the bill at <a href="https://www.kingtuts.co.uk" target="_blank" rel="noopener noreferrer">King Tut's Wah Wah Hut</a> in Glasgow, Scotland.  Alan McGee, the head of Creation Records, is in attendance and offers the band a recording contract on the spot.</span>,
-            date: 'May 31',
-            source: { name: 'Glasgow Live', title: 'Geoff Ellis on King Tut\'s and Remembering that Famous \'93 Oasis Gig', url: 'https://www.glasgowlive.co.uk/whats-on/music-nightlife-news/geoff-ellis-king-tuts-remembering-14712454' },
-            title: 'King Tut\'s Wah Wah Hut, Glasgow, Scotland, UK',
-            type: 'gigs',
-            year: '1993'
-        },
-        {
-            description: <span>At the request of Johnny Marr, Marcus Russell attends an Oasis gig (opening for Dodgy) at Manchester University's Hop & Grape Bar with Marr.  Russell likes the band and can clearly see their influences (Slade, The Sex Pistols and The Kinks).</span>,
-            date: 'June 18',
-            source: { name: 'WalesOnline', title: 'The man who managed Oasis and ran a Welsh ruby club at the same time', url: 'https://www.walesonline.co.uk/sport/rugby/rugby-news/man-who-managed-oasis-ran-16520383.amp'},
-            title: 'Marcus Russell Attends His First Oasis Gig',
-            type: 'noteworthy',
-            year: '1993'
-        },
-        {
-            description: <span>After Marcus Russell attends an Oasis gig the night before in Manchester, Noel Gallagher takes a train down to London and meets with Russell in a cafe in Marlyebone to discuss managing the band.  Noel presents Marcus with a demo tape including <i>Up In The Sky</i>, <i>Whatever</i> and <i>Married With Children</i>.  Within days, Oasis agree to be represented by Russell and his company Ignition Management.</span>,
-            date: 'June 19',
-            source: { name: 'WalesOnline', title: 'The man who managed Oasis and ran a Welsh ruby club at the same time', url: 'https://www.walesonline.co.uk/sport/rugby/rugby-news/man-who-managed-oasis-ran-16520383.amp'},
-            title: 'Noel Gallagher Meets with Marcus Russell in London',
-            type: 'noteworthy',
-            year: '1993'
-        },
-        {
-            description: <span>Recording session at Loco Residential Recording Studios, Newport, Wales, UK.  The session produces demo versions of songs that would later be included on <i>Definitely Maybe</i>.</span>,
-            date: 'September',
-            source: { name: 'Oasis Demo Info', title: 'Definitely Maybe', url: 'https://monobrowdemos.wordpress.com/definitely-maybe-2/' },
-            title: 'Recording Session at Loco Studios, Wales, UK',
-            type: 'recordings',
-            year: '1993'
-        },
-        {
-            description: <span>Oasis play a gig at the Duchess of York pub in Leeds, UK to an empty room after telling the venue manager to charge £50 per ticket as a joke.</span>,
-            date: 'September 8',
-            source: { name: 'Dean Street Press', title: 'Getting High by Paolo Hewitt', url: 'http://deanstreetpress.co.uk/books/hewi02' },
-            title: 'Oasis Play Gig in Leeds to an Empty Room',
-            type: 'gigs',
-            year: '1993'
-        },
-        {
-            description: <span>Broadcast of Oasis' first ever live radio session on the BBC 5 show "Hit The North".  The session was recorded on August 11, 1993.  The band played the following songs: <i>Bring It On Down</i>, <i>I Will Believe</i>, <i>Digsy's Dinner</i>, <i>Cigarettes & Alcohol</i> and <i>Rock 'n' Roll Star</i>.</span>,
-            date: 'September 17',
-            source: { name: 'Oasis Recording Information', title: 'Oasis on the Radio', url: 'http://www.oasis-recordinginfo.co.uk/?page_id=532' },
-            title: 'First Live Radio Broadcast',
-            type: 'noteworthy',
-            year: '1993'
-        },
-        {
-            description: <span><i>Shakermaker</i> recording session at Out of the Blue Studios, Manchester, UK.  This session produced the version that includes the "I'd like to teach the world to sing..." lyrics that would later be re-recorded prior to release as a single.</span>,
-            date: 'October',
-            source: { name: 'Oasis Demo Info', title: 'Definitely Maybe', url: 'https://monobrowdemos.wordpress.com/definitely-maybe-2/' },
-            title: 'Shakermaker Recording Session at Out of the Blue Studios, Manchester, UK',
-            type: 'recordings',
-            year: '1993'
-        },
-        {
-            description: <span>After issues with foreign licensing are resolved, Oasis travel to office of Creation Records in London, UK and sign their first record contract.  They are given and advance of £40,000 and are obliged to release six studio albums with Creation Records via Sony Music.</span>,
-            date: 'October 22',
-            source: { name: 'Dean Street Press', title: 'Getting High by Paolo Hewitt', url: 'http://deanstreetpress.co.uk/books/hewi02' },
-            title: 'Oasis Sign Record Contract with Creation Records',
-            type: 'noteworthy',
-            year: '1993'
-        },
-        {
-            description: <span>First official photo session commissioned by Creation Records with James Fry of the band World of Twist at Nomad Studios in Manchester, UK.  Oasis was heavily influenced by World of Twist and liked their album cover so much that they asked Fry to shoot photos of the band.</span>,
-            date: 'November 29',
-            source: { name: 'British Journal of Photography', title: 'Sons of the Stage: The Story Behind Oasis\'s Rarey Seen Photographs', url: 'https://www.bjp-online.com/2017/10/fullbleed-present-story-behind-oasis-early-photographs/' },
-            title: 'First Official Photo Session with James Fry',
-            type: 'photo',
-            year: '1993'
-        },
-        {
-            description: <span>Creation Records releases a promo 12" 45 RPM vinyl of <i>Columiba</i>.  It is the first Oasis release on Creation and is referred to as the "White Label Demo".</span>,
-            date: 'December',
-            source: { name: 'Discogs', title: 'Columbia White Label Demo', url: 'https://www.discogs.com/Oasis-Columbia/release/757947' },
-            title: 'First Release on Creation Records',
-            type: 'releases',
-            year: '1993'
-        },
-        {
-            description: <span>Phil Smith joins the Oasis road crew for their show at Warwick University.  Smith previously worked as a roadie for the Stone Roses.</span>,
-            date: 'December 8',
-            source: { name: 'Oasis FAQ', title: 'Chronology 1993', url: 'http://www.oocities.org/sunsetstrip/underground/3284/chron93.html'},
-            title: 'Phil Smith Joins Oasis Road Crew',
-            type: 'noteworthy',
-            year: '1993'
-        },
-        {
-            description: <span>What begins as a session to record <i>Bring It On Down</i> as Oasis' first single ends up producing <i>Supersonic</i> at the <a href="https://milocostudios.com/studios/the-motor-museum/intro/" target="_blank" rel="noopener noreferrer">Pink Museum</a> in Liverpool, UK.</span>,
-            date: 'December 19',
-            source: { name: 'Manchester Evening News', title: 'Look Back: Fans Celebrate 20th Anniversary of Debut Oasis Single Supersonic', url: 'https://www.manchestereveningnews.co.uk/news/showbiz-news/look-back-fans-celebrate-music-6432653' },
-            title: 'Supersonic Recording Session at The Pink Museum in Liverpool, UK',
-            type: 'recordings',
-            year: '1993'
-        },
-        {
-            description: <span>Oasis record their first Radio 1 session at Maida Vale Studios in London, UK for <i>The Evening Session</i> program.</span>,
-            date: 'December 22',
-            source: { name: 'Steve Lamacq', title: 'December 1993', url: 'https://www.bbc.co.uk/blogs/stevelamacq/entries/4298f043-9d3c-359f-9178-e1c63ba4c8d4'},
-            title: 'First BBC 1 Radio Session',
-            type: 'noteworthy',
-            year: '1993',
-        },
-        {
-            description: <span>First <i>Definitely Maybe</i> recording session begins at <a href="http://www.monnowvalleystudio.com/" target="_blank" rel="noopener noreferrer">Monnow Valley Residential Recording Studio</a> in Rockfield, Gwentshire, Wales, UK.  The session lasts until January 26, 1994.</span>,
-            date: 'January 7',
-            source: { name: '', title: '', url: '' },
-            title: 'First Definitely Maybe Recording Session at Monnow Valley Studios, Wales, UK',
-            type: 'recordings',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Supersonic</i> on the rooftop of what is now a self storage warehouse between Crestfield and Belgrove Streets in Kings Cross, London, UK.  The video is directed by Mark Szaszy.</span>,
-            date: 'January 24',
-            source: { name: 'Kings Cross Environment', title: 'Oasis in Kings Cross – Supersonic 20 years on', url: 'https://kingscrossenvironment.com/2014/01/24/oasis-in-kings-cross-supersonic-20-years-on/'},
-            title: 'Oasis Film Music Video for Supersonic',
-            type: 'videos',
-            year: '1994'
-
-        },
-        {
-            description: <span>Photoshoot at <a href="http://www.monnowvalleystudio.com/" target="_blank" rel="noopener noreferrer">Monnow Valley Residential Recording Studio</a> in Rockfield, Gwentshire, Wales, UK.  Photos from this session are used for the <i>Supersonic</i> single artwork.</span>,
-            date: 'January 26',
-            source: { name: 'Sonic Editions', title: 'Oasis Supersonic', url: 'https://www.soniceditions.com/image/oasis-at-the-monnow-valley-studios' },
-            title: 'Supersonic Cover Photo Session',
-            type: 'photo',
-            year: '1994'
-        },
-        {
-            description: <span>Widely regarded as Oasis' big breakthrough gig in London and their first gig of 1994, Oasis performs at the legendary <a href="http://www.thewaterratsvenue.london/" target="_blank" rel="noopener noreferrer">Water Rats</a> venue.</span>,
-            date: 'January 27',
-            source: { name: 'Official Oasis Site', title: 'Gigography - Water Rats', url: 'http://www.oasisinet.com/#!/gigs/water-rats' },
-            title: 'Splash Club, Water Rats, Kings Cross, London, UK',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>First gig professionally recorded in full, it was given to a room full of Sony music executives at the Gleneagles Hotel prior to having released any music.  It also represented the start of the UK tour to promote <i>Definitely Maybe</i>. <i>I Am the Walrus</i> is played during the sound check and later released as a b-side to <i>Cigarettes & Alcohol</i>.</span>,
-            date: 'February 6',
-            source: { name: 'Official Oasis Site', title: 'Gigography - Geneagles', url: 'http://www.oasisinet.com/#!/gigs/gleneagles' },
-            title: 'Geneagles Hotel, Pershire, Scotland, UK',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>On their way to Amsterdam, Netherlands to play their first foreign gig in support of The Verve, Liam, Bonhead, Guigsy and Tony are arrested on the ferry ride over and deported.</span>,
-            date: 'February 17',
-            source: { name: 'Live4Ever Forum', title: 'Oasis at the ferry to Amsterdam 1994', url: 'http://live4ever.proboards.com/thread/83015/oasis-ferry-amsterdam-1994' },
-            title: 'Amsterdam Overnight Ferry Incident',
-            type: 'noteworthy',
-            year: '1994'
-        },
-        {
-            description: <span>Second recording session for <i>Definitely Maybe</i> begins at <a href="https://sawmills.co.uk/" target="_blank" rel="noopener noreferrer">Sawmills Residential Recording Studios</a> in Cornwall, UK.  This session would produce most of the tracks for the album.  The session lasts until March 3, 1994.</span>,
-            date: 'February 24',
-            source: { name: '', title: '', url: '' },
-            title: 'Second Definitely Maybe Recording Session at Sawmills Studios in Wales, UK',
-            type: 'recordings',
-            year: '1994'
-        },
-        {
-            description: <span>Mark Coyle and Anjuli Dutt attempt to mix the output of the <i>Definitely Maybe</i> Sawmills Studios recording session at Eden Studios in Chiswick, London, UK.  The result didn't "sound like Oasis", leading the band's manager Marcus Russell to enlist the help of Owen Morris.</span>,
-            date: 'March',
-            source: { name: 'Uncut Magazine', title: 'Oasis\' Definitely Maybe 20 years on...', url: 'https://www.uncut.co.uk/features/oasis-definitely-maybe-20-years-on-8536/' },
-            title: 'Definitely Maybe Mixing Session at Eden Studios in London',
-            type: 'recordings',
-            year: '1994'
-        },
-        {
-            description: <span>Broadcast of Oasis' first ever live TV appearance on Channel 4's The Word.  The band played a shortened version of <i>Supersonic</i>.</span>,
-            date: 'March 18',
-            source: { name: 'Oasis Recording Information', title: 'Oasis on TV', url: 'http://www.oasis-recordinginfo.co.uk/?page_id=455' },
-            title: 'First Live TV Appearance',
-            type: 'noteworthy',
-            year: '1994'
-        },
-        {
-            description: <span>Liam and Noel Gallagher are interviewed by John Harris of the NME at the Forte Crest Hotel in Glasgow, Scotland, UK.  The interview becomes infamous thanks to its release as the bootleg <i>Wibbling Rivalry</i>.</span>,
-            date: 'April 7',
-            source: { name: 'Oasis Interviews', title: 'Noel & Liam Gallager Interview - April 7th 1994', url: 'http://oasisinterviews.blogspot.com/1994/04/noel-liam-gallagher-wibbling-rivalry.html' },
-            title: 'John Harris Interviews Liam and Noel Gallagher',
-            type: 'noteworthy',
-            year: '1994'
-        },
-        {
-            description: <span><i>Supersonic</i>, the first single from the forthcoming album <i>Definitely Maybe</i> is released by Creation Records.</span>,
-            date: 'April 11',
-            source: { name: 'Wikipedia', title: 'Supersonic (Oasis song)', url: 'https://en.wikipedia.org/wiki/Supersonic_(Oasis_song)' },
-            title: 'Supersonic Released',
-            type: 'releases',
-            year: '1994'
-        },
-        {
-            description: <span>The music video for <i>Supersonic</i> is released.  The video is directed by Mark Szaszy.</span>,
-            date: 'April 11',
-            source: { name: 'IMDB', title: 'Oasis: Supersonic', url: 'https://www.imdb.com/title/tt5685696/' },
-            title: 'Supersonic Music Video Released',
-            type: 'videos',
-            year: '1994'
-        },
-        {
-            description: <span>To celebrate the release of their debut single, <i>Supersonic</i>, the band take over the Wheatsheaf pub in Stoke-on-Trent, UK.  Despite discovering the band less than a year earlier, Alan McGee is not in attendance due to checking himself into rehab for drug addiction.</span>,
-            date: 'April 11',
-            source: { name: 'StokeonTrentLive', title: '\'Rehab made me miss Supersonic launch in Stoke\', says Alan McGee - the man who discovered Oasis', url: 'https://www.stokesentinel.co.uk/whats-on/rehab-made-miss-supersonic-launch-2772047' },
-            title: 'Oasis Celebrate First Single Release in Stoke',
-            type: 'noteworthy',
-            year: '1994'
-        },
-        {
-            description: <span><i>Cigarettes & Alcohol</i> single cover photo session at Halcyon Hotel in London, UK.  Artwork features members of the band plus Tim Abbott (managing director of Creation Records at the time), Jane Fisher and Emma Evans of Creation Records.</span>,
-            date: 'August 14',
-            source: { name: 'Sonic Editions', title: 'Cigarettes and Alcohol Front Cover', url: 'https://www.soniceditions.com/image/cigarettes-and-alcohol-front-cover' },
-            title: 'Cigarettes & Alcohol Cover Photo Session',
-            type: 'photo',
-            year: '1994'
-        },
-        {
-            description: <span><i>Definitely Maybe</i> album cover photo session at Paul "Bonehead" Arthur's house in Didsbury, Manchester, UK.  The cover was designed by Brian Cannon of Microdot Creative and shot by Michael Spencer Jones.</span>,
-            date: 'May 30',
-            source: { name: 'Sonic Editions', title: 'Oasis Definitely Maybe', url: 'https://www.soniceditions.com/image/oasis-definitely-maybe' },
-            title: 'Definitely Maybe Cover Photo Session',
-            type: 'photo',
-            year: '1994'
-        },
-        {
-            description: <span>First appearance on the cover of the New Musical Expression magazine.  Liam is pictured standing in front of The Oasis Bar with the caption: "Totally Cool - Oasis: What the world is waiting for".</span>,
-            date: 'June 4',
-            source: { name: 'Official Oasis Site', title: 'Press Archive - NME 04.06.1994', url: 'http://www.oasisinet.com/#!/press/covers/totally-cool-nme-04-06-1994' },
-            title: 'First NME Cover',
-            type: 'noteworthy',
-            year: '1994'
-        },
-        {
-            description: <span><i>Shakermaker</i>, the second single from <i>Definitely Maybe</i>, is released by Creation Records.</span>,
-            date: 'June 13',
-            source: { name: 'Wikipedia', title: 'Shakermaker', url: 'https://en.wikipedia.org/wiki/Shakermaker' },
-            title: 'Shakermaker Released',
-            type: 'releases',
-            year: '1994'
-        },
-        {
-            description: <span>The music video for <i>Shakermaker</i> is released.  The video is directed by Mark Szaszy.</span>,
-            date: 'June 13',
-            source: { name: 'IMDB', title: 'Oasis: Shakermaker', url: 'https://www.imdb.com/title/tt5685720/' },
-            title: 'Shakermaker Music Video Released',
-            type: 'videos',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis play their first ever international gig at Erotika in Paris, France.</span>,
-            date: 'June 16',
-            source: { name: 'This Day In Rock', title: 'This Day In Rock', url: 'http://www.thisdayinrock.com/index.php/general/1994-oasis-played-at-the-erotika-club-paris-in-france-the-band/' },
-            title: 'Erotika, Paris, France',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>First appearance at the Glastonbury Festival.</span>,
-            date: 'June 26',
-            source: { name: 'The NME', title: '9 Things We Learned From This Recently Unearthed Footage of Oasis\' First Ever Glastonbury Appearance', url: 'https://www.nme.com/blogs/festivals-blog/oasis-first-ever-glastonbury-appearance-11231' },
-            title: 'Glastonbury Festival, Worthy Farm, Shepton Mallet, Pilton, UK',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis records an appearance on the BBC TV show Top of the Pops for the first time.  The band "performs" <i>Shakermaker</i>.  The show is broadcast on June 30, 1994.</span>,
-            date: 'June 29',
-            source: { name: 'Turn Up The Volume Blog', title: 'Debut on Top of the Pops', url: 'https://turnupthevolume.blog/2015/06/29/oasis-june-29-1994-debut-on-top-of-the-pops/' },
-            title: 'First Appearance on Top of the Pops',
-            type: 'noteworthy',
-            year: '1994'
-        },
-        {
-            description: <span>Recording session at Maison Rouge in Fulham, London, UK.  Versions of <i>Whatever</i>, <i>Listen Up</i>, <i>Fade Away</i> and a demo version of <i>Some Might Say</i> are recorded.</span>,
-            date: 'July 6',
-            source: { name: 'Oasis Interviews', title: 'Noel & Liam Gallagher Interview - Melody Maker April 27 1996', url: 'http://oasisinterviews.blogspot.com/1996/04' },
-            title: 'Whatever Recording Session at Maison Rouge',
-            type: 'recordings',
-            year: '1994'
-        },
-        {
-            description: <span>Whilst recording at Maison Rouge in Fulham, London, UK, band members trash the bar and rooms at the Columbia Hotel in Paddington, London, UK.  Some furniture tossed from the room lands on the car belonging to the hotel manager.  This leads to the band members being banned for life from the hotel.</span>,
-            date: 'July 6',
-            source: { name: 'Oasis Interviews', title: 'Noel & Liam Gallagher Interview - Melody Maker April 27 1996', url: 'http://oasisinterviews.blogspot.com/1996/04' },
-            title: 'Oasis Banned from Columbia Hotel in London, UK',
-            type: 'recordings',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis play their first gig in North America at the Wetlands Preserve Festival in New York, New York, USA.</span>,
-            date: 'July 21',
-            source: { name: 'The Telegraph', title: 'Paul \'Bonehead\' Arthurs Remembers Oasis\'s First American Tour, 1994', url: 'https://www.telegraph.co.uk/culture/music/rockandpopmusic/10741593/Paul-Bonehead-Arthurs-remembers-Oasiss-first-American-tour-1994.html' },
-            title: 'Wetlands Preserve Festival, New York, New York, USA',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis film the UK version of the music video for <i>Live Forever</i> in Central Park, New York City, New York, USA.  Locations used include Starwberry Fields and the Naumburg Bandshell.  The video is directed by Carlos Grasso.</span>,
-            date: 'July 24',
-            source: { name: 'Live4Ever Forum', title: 'Oasis\' Videos', url: 'http://live4ever.proboards.com/post/393777' },
-            title: 'Live Forever (UK Version) Music Video Filmed',
-            type: 'videos',
-            year: '1994'
-        },
-        {
-            description: <span><i>Live Forever</i>, the third single from <i>Definitely Maybe</i>, is released by Creation Records.</span>,
-            date: 'August 8',
-            source: { name: 'Wikipedia', title: 'Live Forever (Oasis song)', url: 'https://en.wikipedia.org/wiki/Live_Forever_(Oasis_song)' },
-            title: 'Live Forever Released',
-            type: 'releases',
-            year: '1994'
-        },
-        {
-            description: <span>The UK version of the music video for <i>Live Forever</i> is released.  The video is directed by Carlos Grasso.</span>,
-            date: 'August 8',
-            source: { name: 'IMDB', title: 'Oasis: Live Forever', url: 'https://www.imdb.com/title/tt5685656/' },
-            title: 'Live Forever (UK Version) Music Video Released',
-            type: 'videos',
-            year: '1994'
-        },
-        {
-            description: <span>After back and forth with fans during a gig at Riverside in Newcastle, UK, a fight breaks out on stage resulting in Noel Gallager damaging a 1960's sunburst Gibson Les Paul guitar given to him by Johnny Marr of The Smiths.  The band refuse to continue the show after 5 songs, leading to fans surrounding the band's van.  Noel also would require stitches after the attack.</span>,
-            date: 'August 9',
-            source: { name: 'Chronicle Live', title: 'Brawl breaks out at Oasis gig at the Riverside in Newcastle in 1994', url: 'https://www.chroniclelive.co.uk/news/history/brawl-breaks-out-oasis-gig-11721778' },
-            title: 'Oasis Brawl with Fans at Riverside in Newcastle, UK',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>After playing a set at the Hultsfred Festival in Stockholm, Sweden, Oasis, The Verve and Primal Scream cause over £1,000 of damage to a hotel bar.  The band is arrested and asked to leave Sweden.</span>,
-            date: 'August 13',
-            source: { name: 'exclaim!*@#', title: 'Oasis (Here\'s The Story) Morning Glory', url: 'http://exclaim.ca/music/article/oasis-_heres_the_story_morning_glory' },
-            title: 'Oasis Play Hultsfred Festival/Labeled "English Scum" by Swedish Press',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis record the music video for <i>Cigarettes & Alcohol</i> at the Borderline Club, London, UK.  After filming for the video, Oasis play an 11 song set to the 500 guests/extras in attendance for the video shoot.</span>,
-            date: 'August 21',
-            source: { name: 'Oasis Bootlegs From Fuckuoka', title: 'Was There Then', url: 'https://oasisfromfuckuoka.blogspot.com/2012/01/was-there-then-flbb038.html' },
-            title: 'Cigarettes & Alcohol Music Video Filmed',
-            type: 'videos',
-            year: '1994'
-        },
-        {
-            description: <span><i>Definitely Maybe</i>, Oasis' debut studio album, is released by Creation Records.  It would go on to become the fastest selling debut album in UK chart history at the time.</span>,
-            date: 'August 29',
-            source: { name: 'Wikipedia', title: 'Definitely Maybe', url: 'https://en.wikipedia.org/wiki/Definitely_Maybe' },
-            title: 'Definitely Maybe Released',
-            type: 'releases',
-            year: '1994'
-        },
-        {
-            description: <span>To celebrate the release of <i>Definitely Maybe</i>, Oasis play a short set of six songs at London's Virgin Megastore.  Evan Dando (of the band The Lemonheads) joins Oasis on stage, playing the tambourine.</span>,
-            date: 'August 29',
-            source: { name: 'Music History Calendar', title: 'Oasis Starts Britpop Invasion With Definitely Maybe', url: 'https://calendar.songfacts.com/august/29/17571' },
-            title: 'Oasis Play the Virgin Metastore in London to Promote the Release of Definitely Maybe',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Definitely Maybe</i> album sales as gold.</span>,
-            date: 'September 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Definitely Maybe Sales Certified as Gold by BPI',
-            type: 'certifications',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis perform their first gig in Japan at the Quattro in Tokyo.</span>,
-            date: 'September 13',
-            source: { name: 'Oasis Official Site', title: 'Gigography - 1994', url: 'http://www.oasisinet.com/#!/gigs/list?filter_date=1994' },
-            title: 'Quattro, Tokyo, Japan',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis perform an infamous gig at the Whiskey A Go Go in Los Angeles, California, USA that results in Noel Gallagher leaving the band.  He would rejoin them about a week later and continue the tour.</span>,
-            date: 'September 29',
-            source: { name: 'Oasis Official Site', title: 'Gigography - 1994', url: 'http://www.oasisinet.com/#!/gigs/list?filter_date=1994' },
-            title: 'The Whiskey A Go Go, Los Angeles, California, USA',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis film the US version of the music video <i>Supersonic</i> at the <a href="https://www.cabazondinosaurs.com/" target="_blank" rel="noopener noreferrer">Cabazon Dinosaurs</a> in Cabazon, California, USA.  The video is directed by Nick Egan.</span>,
-            date: 'October ?',
-            source: { },
-            title: 'Oasis Film US Version of Supersonic Music Video',
-            type: 'videos',
-            year: '1994'
-        },
-        {
-            description: <span>Recording session at the <a href="http://congresshouse.com/" target="_blank" rel="noopener noreferrer">Congress House Studio</a> in Austin, TX while on tour in the US during September/October, 1994.</span>,
-            date: 'October 8',
-            source: { name: 'Oasis Recording Information', title: 'The Recording Sessions', url: 'http://www.oasis-recordinginfo.co.uk/?page_id=702' },
-            title: 'Recording Session at Congress House, Austin, Texas, USA',
-            type: 'recordings',
-            year: '1994'
-        },
-        {
-            description: <span><i>Cigarettes & Alcohol</i>, the fourth single from <i>Definitely Maybe</i>, is released by Creation Records.</span>,
-            date: 'October 10',
-            source: { name: 'Wikipedia', title: 'Cigarettes & Alcohol', url: 'https://en.wikipedia.org/wiki/Cigarettes_%26_Alcohol' },
-            title: 'Cigarettes & Alcohol Released',
-            type: 'releases',
-            year: '1994'
-        },
-        {
-            description: <span>The music video for <i>Cigarettes & Alcohol</i> is released.  The video is directed by Mark Szaszy.</span>,
-            date: 'October 10',
-            source: { name: 'IMDB', title: 'Oasis: Cigarettes & Alcohol', url: 'https://www.imdb.com/title/tt5685656/' },
-            title: 'Cigarettes & Alcohol Music Video Released',
-            type: 'videos',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis perform at the Cabaret Metro in Chicago, Illinois, USA.  The gig is the first professionally recorded show for radio broadcast in North America.  It is made available as a <a href="https://www.discogs.com/Oasis-Live/release/1622979" target="_blank" rel="noopener noreferrer">promotional release</a> for US radio stations.</span>,
-            date: 'October 15',
-            source: { name: 'Official Oasis Site', title: 'Gigography - Chicago 1994', url: 'http://www.oasisinet.com/#!/gigs/cabaret-metro-chicago-1994' },
-            title: 'Cabaret Metro, Chicago, Illinois, USA',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>Photo session in Derbyshire, UK to produce the cover art for the upcoming <i>Whatever</i> single release.</span>,
-            date: 'October 18',
-            source: { name: 'Sonic Editions', title: 'Oasis Whatever', url: 'https://www.soniceditions.com/image/oasis-whatever' },
-            title: 'Whatever Cover Photo Session',
-            type: 'photo',
-            year: '1994'
-        },
-        {
-            description: <span>Recording session for the <i>Whatever</i> single at <a href="http://www.rockfieldmusicgroup.com/default.asp?contentID=543" target="_blank" rel="noopener noreferrer">Rockfield Studios</a> in Gwent, Wales, UK.</span>,
-            date: 'November',
-            source: { name: 'Wikipedia', title: 'Whatever (Oasis song)', url: 'https://en.wikipedia.org/wiki/Whatever_(Oasis_song)' },
-            title: 'Whatever Recording Session',
-            type: 'recordings',
-            year: '1994'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Definitely Maybe</i> album sales as platinum.</span>,
-            date: 'December 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Definitely Maybe Sales Certified as Platinum by BPI',
-            type: 'certifications',
-            year: '1994'
-        },
-        {
-            description: <span>While playing at the Barrowland Ballroom in Glasgow, Scotland, UK, Liam Gallagher loses his voice.  After a few hectic moments and boos, Noel Gallagher returns to the stage and plays a seven song acoustic set to calm the crowd.</span>,
-            date: 'December 7',
-            source: { name: 'The Herald', title: 'Oasis, Barrowland, Glasgow A result in extra time', url: 'https://www.heraldscotland.com/news/12542671.oasis-barrowland-glasgow-a-result-in-extra-time/' },
-            title: 'Liam Gallagher Loses His Voice/Noel Gallagher Plays Acoustic Solo Set',
-            type: 'gigs',
-            year: '1994'
-        },
-        {
-            description: <span>Oasis record the video for <i>Whatever</i> in Wimbledon, London, UK.  Noel Gallagher would later admit that he had been out the entire night before and couldn't remember the location of the video shoot.  The video is directed by Mark Szaszy.</span>,
-            date: 'December 12',
-            source: { name: 'Live4Ever Forum', title: 'Oasis\' Videos', url: 'http://live4ever.proboards.com/post/393777/thread' },
-            title: 'Whatever Music Video Filmed',
-            type: 'videos',
-            year: '1994'
-        },
-        {
-            description: <span><i>Whatever</i>, a stand-alone single between the albums <i>Definitely Maybe</i> and <i>(What's the Story) Morning Glory?</i> is released.</span>,
-            date: 'December 18',
-            source: { name: 'Wikipedia', title: 'Whatever (Oasis song)', url: 'https://en.wikipedia.org/wiki/Whatever_(Oasis_song)' },
-            title: 'Whatever Released',
-            type: 'releases',
-            year: '1994'
-        },
-        {
-            description: <span>The music video for <i>Whatever</i> is released.  The video is directed by Mark Szaszy.</span>,
-            date: 'February',
-            source: { name: 'IMDB', title: 'Oasis: Whatever', url: 'https://www.imdb.com/title/tt5685668/' },
-            title: 'Whatever Music Video Released',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span>The US version of the music video for <i>Live Forever</i> is released.  The video is directed by Nick Egan.</span>,
-            date: 'February',
-            source: { name: 'IMDB', title: 'Oasis: Live Forever', url: 'https://www.imdb.com/title/tt6023796/' },
-            title: 'Live Forever (US Version) Music Video Released',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span>Recording session that produces <i>Some Might Say</i>, <i>Acquiesce</i> and <i>Headshrinker</i>.</span>,
-            date: 'March',
-            source: { name: 'Sound on Sound', title: 'Oasis \'Wonderwall\'', url: 'https://www.soundonsound.com/people/oasis-wonderwall' },
-            title: 'Some Might Say Recording Session at Loco Studios, Newport, Wales, UK',
-            type: 'recordings',
-            year: '1995'
-        },
-        {
-            description: <span>Filming of the music video is booked to begin at Chatley Heath, Surrey, UK.  It has a budget of £40,000 and is to last two days.  However, Liam Gallagher refuses to turn up, causing the video shoot to be cancelled.  Instead, a video using clips from other Oasis music videos (<i>Cigarettes & Alcohol</i>, the US version of <i>Supersonic</i> and <i>Whatever</i>) is produced.</span>,
-            date: 'March 29',
-            source: { name: 'Dean Street Press', title: 'Getting High by Paolo Hewitt', url: 'http://deanstreetpress.co.uk/books/hewi02' },
-            title: 'Some Might Say Music Video Shoot Cancelled',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Definitely Maybe</i> album sales as 2x platinum.</span>,
-            date: 'April 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Definitely Maybe Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis record a performance for Channel 4's <i>The White Room</i>.  Instead of playing the upcoming single <i>Some Might Say</i>, the band opt instead to play <i>Acquiesce</i>, <i>It's Good To Be Free</i> and <i>Talk Tonight</i>, the latter featuring Paul Weller on electric piano.</span>,
-            date: 'April 14',
-            source: { name: 'Louder Than War', title: 'Oasis: Some Might Say - 20th Anniversary', url: 'https://louderthanwar.com/might-say-number-1-20th-anniversary/' },
-            title: 'Oasis Record Live Performance for The White Room',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Rock 'n' Roll Star</i> at the amusement park at the Southend Pier.  The final cut also includes footage from Oasis' gig at the Cliffs Pavilion.  The video is directed by Nigel Dick.</span>,
-            date: 'April 16',
-            source: { name: 'Nigel Dick Official Site', title: 'Productions 1995', url: 'http://www.nigeldick.com/films/productions-1995/' },
-            title: 'Rock \'n\' Roll Star Music Video Filmed',
-            type: 'gigs',
-            year: '1995'
-        },
-        {
-            description: <span>Concert released as the official video <i>Live By The Sea</i></span>,
-            date: 'April 17',
-            source: { name: 'Official Oasis Site', title: 'Gigography - Cliffs Pavilion 1995', url: 'http://www.oasisinet.com/#!/gigs/cliffs-pavillion-1995' },
-            title: 'Cliffs Pavilion, Southend, UK',
-            type: 'gigs',
-            year: '1995'
-        },
-        {
-            description: <span>The last show of the <i>Definitely Maybe</i> tour and Oasis' first arena gig.  It would be Tony McCarroll's last live performance with the band.  It is also noteworthy for including the debut of <i>Don't Look Back In Anger</i> performed acoustically by Noel Gallagher.</span>,
-            date: 'April 22',
-            source: { name: 'Louder Than War', title: 'Oasis: Some Might Say - 20th Anniversary', url: 'https://louderthanwar.com/might-say-number-1-20th-anniversary/' },
-            title: 'Sheffield Arena, Sheffield, UK',
-            type: 'gigs',
-            year: '1995'
-        },
-        {
-            description: <span><i>Some Might Say</i>, the first single from the forthcoming album <i>(What's The Story) Morning Glory?</i>, is released on Creation Records.  It would go on to become Oasis' first number one single and is the last recording to feature the original lineup.  The single includes the b-sides <i>Talk Tonight</i>, <i>Acquiesce</i> and <i>Headshrinker</i>.</span>,
-            date: 'April 24',
-            source: { name: 'Wikipedia', title: 'Some Might Say', url: 'https://en.wikipedia.org/wiki/Some_Might_Say' },
-            title: 'Some Might Say Released',
-            type: 'releases',
-            year: '1995'
-        },
-        {
-            description: <span>The music video for <i>Some Might Say</i> is released.  The video is directed by Stuart Fryer.</span>,
-            date: 'April 24',
-            source: { name: 'IMDB', title: 'Oasis: Some Might Say', url: 'https://www.imdb.com/title/tt7085562/' },
-            title: 'Some Might Say Music Video Released',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span>Tony McCarroll records what will be his last public appearance with the band during a live "performance" of <i>Some Might Say</i> on the BBC program <i>Top of the Pops</i>.</span>,
-            date: 'April 26, 1995',
-            source: { name: 'Louder Than War', title: 'Oasis: Some Might Say - 20th Anniversary', url: 'https://louderthanwar.com/might-say-number-1-20th-anniversary/' },
-            title: 'Tony McCarroll\'s Last Appearance as Drummer',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis announces that Tony McCarroll has left the band.</span>,
-            date: 'April 30',
-            source: { name: 'Official Oasis Site', title: 'Oasis Lose Drummer Tony McCarroll', url: 'https://web.archive.org/web/19970605053939/http://www.oasisinet.com/tonyalan.htm' },
-            title: 'Tony McCarroll Leaves Oasis',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>The music video for <i>Rock 'n' Roll Star</i> is released.  The video is directed by Nigel Dick.</span>,
-            date: 'May',
-            source: { name: 'IMDB', title: 'Oasis: Rock \'n\' Roll Star', url: 'https://www.imdb.com/title/tt7085562/' },
-            title: 'Rock \'n\' Roll Star Music Video Released',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis announce that Alan White has joined Oasis to replace the recently departed Tony McCarroll as the band's full time drummer.</span>,
-            date: 'May 2',
-            source: { name: 'Official Oasis Site', title: 'Oasis Hire Alan White', url: 'https://web.archive.org/web/19970605053939/http://www.oasisinet.com/tonyalan.htm' },
-            title: 'Alan White Announced as Replacement Drummer',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>Alan White records his first public appearance with the band during a live "performance" of <i>Some Might Say</i> on the BBC program <i>Top of the Pops</i>.</span>,
-            date: 'May 3',
-            source: { name: 'Louder Than War', title: 'Oasis: Some Might Say - 20th Anniversary', url: 'https://louderthanwar.com/might-say-number-1-20th-anniversary/' },
-            title: 'Alan White\'s First Appearance as Drummer',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>Recording session for Oasis' second album <i>(What's The Story) Morning Glory?</i> at <a href="http://www.rockfieldmusicgroup.com/default.asp?contentID=543" target="_blank" rel="noopener noreferrer">Rockfield Studios</a> in Gwent, Wales, UK.  The session was famously interrupted in a brawl in which Noel struck Liam over the head with a cricket bat.  It would resume in June after tempers had subsided.</span>,
-            date: 'May 8',
-            source: { name: 'Sound on Sound', title: 'Oasis \'Wonderwall\'', url: 'https://www.soundonsound.com/people/oasis-wonderwall' },
-            title: '(What\'s The Story) Morning Glory Recording Session - Rockfield Studios, Gwent, Wales UK',
-            type: 'recordings',
-            year: '1995'
-        },
-        {
-            description: <span>After a trip to a pub in Monmouth, Liam Gallagher returns to Rockfield Studio with journalist John Robb, members of the band Cable and some local residents.  An altercation broke out between members of Cable and Liam in the residence at the studio.  Noel Gallagher confronts brother Liam about the fracas, starting an argument that leads to Liam breaking his foot while kicking in Noel's bedroom door and Noel striking Liam with a cricket bat.  The result was the abrupt end of the recording session after only one week.</span>,
-            date: 'May 12',
-            source: { name: 'Sound on Sound', title: 'Oasis \'Wonderwall\'', url: 'https://www.soundonsound.com/people/oasis-wonderwall' },
-            title: 'Brawl ends (What\'s The Story) Morning Glory Recording Session',
-            type: 'noteworty',
-            year: '1995'
-        },
-        {
-            description: <span>After the previous session ended in an intra-band fight, Oasis resumes recording its second album <i>(What's The Story) Morning Glory?</i> at <a href="http://www.rockfieldmusicgroup.com/default.asp?contentID=543" target="_blank" rel="noopener noreferrer">Rockfield Studios</a> in Gwent, Wales, UK.</span>,
-            date: 'May 22',
-            source: { name: 'Sound on Sound', title: 'Oasis \'Wonderwall\'', url: 'https://www.soundonsound.com/people/oasis-wonderwall' },
-            title: '(What\'s The Story) Morning Glory Recording Session Resumes at Rockfield Studios, Gwent, Wales, UK',
-            type: 'recordings',
-            year: '1995'
-        },
-        {
-            description: <span><i>Roll With It</i> cover photo session in Weston-super-Mare, UK.</span>,
-            date: 'June 22',
-            source: { name: 'Sonic Editions', title: 'Oasis Roll With It', url: 'https://www.soniceditions.com/image/oasis-roll-with-it' },
-            title: 'Roll With It Cover Photo Session',
-            type: 'photo',
-            year: '1995'
-        },
-        {
-            description: <span>First gig of the world tour to support Oasis' second album <i>(What's The Story) Morning Glory?</i> at Bath Pavilion, Bath, UK.  It is the first gig with Alan White as drummer and is also the live debut of the songs <i>Hello</i>, <i>Roll With It</i> and <i>Morning Glory</i>.</span>,
-            date: 'June 22',
-            source: { name: 'Wikipedia', title: '(What\'s the Story) Morning Glory?', url: 'https://en.wikipedia.org/wiki/(What%27s_the_Story)_Morning_Glory%3F_Tour' },
-            title: 'First Gig of the (What\'s The Story) Morning Glory? World Tour',
-            type: 'gigs',
-            year: '1995'
-        },
-        {
-            description: <span>The Recording Industry Association of America certifies <i>Definitely Maybe</i> album sales as gold.</span>,
-            date: 'June 22',
-            source: { name: 'Recording Industry Association of America', title: 'Gold & Platinum Awards', url: 'https://www.riaa.com/gold-platinum/?tab_active=default-award&ar=Oasis&ti=&lab=&genre=&format=&date_option=release&from=1993-01-01&to=1998-12-31&award=&type=&category=&adv=SEARCH#search_section' },
-            title: 'Definitely Maybe Sales Certified as Gold by RIAA',
-            type: 'certifications',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis headline the first night of the Glastonbury Festival at Worthy Farm in Pilton, UK.  They split their set between songs from <i>Definitely Maybe</i> and new songs from their upcoming second album <i>(What's the Story) Morning Glory?</i>.</span>,
-            date: 'June 23',
-            source: { name: 'Official Oasis Site', title: 'Gigography - Glastonbury Festival', url: 'http://www.oasisinet.com/#!/gigs/glastonbury-festival' },
-            title: 'Glastonbury Festival, Worthy Farm, Shepton Mallet, Pilton, UK',
-            type: 'gigs',
-            year: '1995'
-        },
-        {
-            description: <span>Citing hearing loss, Mark Coyle works his last gig at Slane Castle in Ireland (opening for R.E.M.) as the front of house engineer for Oasis.</span>,
-            date: 'July 22',
-            source: { name: 'The Irish Post', title: 'Oasis were different', url: 'https://www.irishpost.com/entertainment/oasis-different-like-mates-storming-palace-34219' },
-            title: 'Mark Coyle Quits as Front of House Engineer',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis perform live at a studio in Kings Cross, London, UK to fans who won tickets via a phone-in competition.  After filming the portion for the music video, Oasis plays a full gig for the fans.  The video is directed by John Klein.</span>,
-            date: 'July 26',
-            source: { name: 'Dean Street Press', title: 'Getting High by Paolo Hewitt', url: 'http://deanstreetpress.co.uk/books/hewi02' },
-            title: 'Oasis Film Roll With It Music Video',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span><i>(What's The Story) Morning Glory?</i> cover photo session on Berwick Street, Soho, London,UK.  The photo features Brian Cannon (who designed all previous covers) walking towards DJ Sean Rowley.  You can also spot Owen Morris (producer) in the background holding the album master tapes over his face.</span>,
-            date: 'July 27',
-            source: { name: 'Sonic Editions', title: 'Oasis Morning Glory', url: 'https://www.soniceditions.com/image/oasis-morning-glory' },
-            title: '(What\'s The Story) Morning Glory? Cover Photo Session',
-            type: 'photo',
-            year: '1995'
-        },
-        {
-            description: <span><i>Wonderwall</i> cover photo session in Primrose Hill, London, UK.  Initial photographs from this session had Liam inside the picture frame.  The final picture features Anita Heyret, an employee of Creation Records.</span>,
-            date: 'August 1',
-            source: { name: 'Sonic Editions', title: 'Oasis Wonderwall', url: 'https://www.soniceditions.com/image/oasis-wonderwall' },
-            title: 'Wonderwall Cover Photo Session',
-            type: 'photo',
-            year: '1995'
-        },
-        {
-            description: <span>Paul "Bonehead" Arthurs marries his childhood girlfriend Kate at the registry office in Manchester, UK.</span>,
-            date: 'August 10',
-            source: { name: 'Official Oasis Site', title: 'Bonehead Marries Sweetheart', url: 'https://web.archive.org/web/19970605053924/http://www.oasisinet.com/mrshead.htm' },
-            title: 'Paul "Bonehead" Arthurs Marries Girlfriend Kate',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span><i>Roll With It</i>, the second single from the album <i>(What's The Story) Morning Glory?</i>, is released on Creation Records.  Famously came in second to Blur's <i>Country House</i> in a battle of the charts.  The single includes the b-sides <i>It's Better People</i>, <i>Rockin' Chair</i> and <i>Live Forever</i> live at Glastonbury, 1995.</span>,
-            date: 'August 14',
-            source: { name: 'Wikipedia', title: 'Roll With It (Oasis song)', url: 'https://en.wikipedia.org/wiki/Roll_with_It_(Oasis_song)' },
-            title: 'Roll With It Released',
-            type: 'releases',
-            year: '1995'
-        },
-        {
-            description: <span>The music video for <i>Roll With It</i> is released.  The video is directed by Jon Klein.</span>,
-            date: 'August 14',
-            source: { name: 'IMDB', title: 'Oasis: Roll With It', url: 'https://www.imdb.com/title/tt5638926/' },
-            title: 'Roll With It Music Video Released',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span>After a week of sales, it is revealed that Blur's <i>Country House</i> has beaten Oasis' <i>Roll With It</i> to number on in the UK singles charts.  Blur appear on the BBC's <i>Top of the Pops</i> that week with bassist Alex James wearing an Oasis T-Shirt.</span>,
-            date: 'August 20',
-            source: { name: 'BBC World News', title: 'Timeline: Blur v Oasis after Britpop', url: 'http://news.bbc.co.uk/2/hi/entertainment/4151510.stm' },
-            title: 'Blur Beat Oasis to Number One',
-            type: 'releases',
-            year: '1995'
-        },
-        {
-            description: <span>Officially released live video of Oasis performing at Cliffs Pavilion, Southend, UK on April 17, 1995.</span>,
-            date: 'August 28',
-            source: { name: 'Wikipedia', title: 'Live by the Sea', url: 'https://en.wikipedia.org/wiki/Live_by_the_Sea' },
-            title: 'Live By The Sea Released',
-            type: 'releases',
-            year: '1995'
-        },
-        {
-            description: <span>Noel Gallagher records a cover of the Beatles' <i>Come Together</i> with Paul McCartney and Paul Weller at <a href="https://www.abbeyroad.com/" target="_blank" rel="noopener noreferrer">Abbey Road Studios</a> in London, UK for War Child's <i>Help</i> album.</span>,
-            date: 'September 4',
-            source: { name: 'Q Magazine', title: 'Celebrating War Child\'s HELP Album 20 Years On', url: 'https://www.qthemusic.com/articles/column/guest-column-war-child-help-album' },
-            title: 'Noel Gallagher Records with Paul McCartney and Paul Weller',
-            type: 'recordings',
-            year: '1995'
-        },
-        {
-            description: <span>Bassist Paul "Guigsy" McGuigan leaves the band due to nervous exhaustion from touring, leading Oasis to cancel its upcoming tour dates in the US.  While unconfirmed, it is believed that an altercation in Paris while on vacation with Liam eventually leads to Guigsy bowing out of the band temporarily.  Scott McLeod of The Ya Ya's joins Oasis as their new bassit.</span>,
-            date: 'September 12',
-            source: { name: 'Official Oasis Site', title: 'Oasis Postpone September Tour Dates', url: 'https://web.archive.org/web/19970605053915/http://www.oasisinet.com/guigill.htm' },
-            title: 'Guigsy Leaves Oasis',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>In an article published in The Observer (The Guardian), Noel Gallagher says of his rivarly with Blur: "I hope the pair of them catch AIDS and die because I fucking hate them two" (referring to Damon Albarn and Alex James).  Noel immediately retracts the comment, saying that AIDS is no joking matter.</span>,
-            date: 'September 17',
-            source: { name: 'Oasis Interviews', title: 'Noel Gallgher Interview September 17, 1995 - The Guardian', url: 'http://oasisinterviews.blogspot.com/1995/09/noel-gallagher-guardian-17th-september.html' },
-            title: 'Noel Gallagher Wishes Damon Albarn and Alex James of Blur die of AIDS',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>After Paul "Guigsy" McGuigan quit touring due to exhaustion, Oasis draft Scott McLeod of The Ya-Yas to fill in as their temporary bass guitarist for their live tour commitments.</span>,
-            date: 'September 26',
-            source: { name: 'Official Oasis Site', title: 'Oasis Name Temporary Live Bassist', url: 'https://web.archive.org/web/19970605053859/http://www.oasisinet.com/tempbass.htm' },
-            title: 'Scott McLeod Joins Oasis as Temporary Live Bassist',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Wonderwall</i> at Unit 217B in Woolwich, London, UK.  The video includes Scott McLeod, who joined Oasis a few days earlier as a replacement for Paul "Guigsy" McGuigan.  The video is directed by Nigel Dick.</span>,
-            date: 'September 30',
-            source: { name: 'Nigel Dick Official Site', title: 'Productions 1995', url: 'http://www.nigeldick.com/films/productions-1995/' },
-            title: 'Wonderwall Music Video Filmed',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis holds a launch party for <i>(What's The Story) Morning Glory?</i> in Kensington, London, UK.  All guess receive an <a href="https://www.alessi.com/us_en/product/kitchen/kitchen-boxes-biscuit-boxes-and-containers/kitchen-box-with-hermetic-lid-gianni-a-little-man-holding-on-tight-pc-amdr04.html" target="_blank" rel="noopener noreferrer">Allessi Jar</a> with a cassette copy of the album.</span>,
-            date: 'October 1',
-            source: { name: 'Dean Street Press', title: 'Getting High by Paolo Hewitt', url: 'http://deanstreetpress.co.uk/books/hewi02' },
-            title: '(What\'s The Story) Morning Glory? Launch Party',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 2x platinum.</span>,
-            date: 'October 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '1995'
-        },
-        {
-            description: <span><i>(What's the Story) Morning Glory?</i>, Oasis' second studio album, is released on Creation Records.  The album would propel the band to a worldwide fame, selling over 12 million copies around the world.</span>,
-            date: 'October 2',
-            source: { name: 'Wikipedia', title: '(What\'s the Story) Morning Glory?', url: 'https://en.wikipedia.org/wiki/(What%27s_the_Story)_Morning_Glory%3F' },
-            title: '(What\'s The Story) Morning Glory? Released',
-            type: 'releases',
-            year: '1995'
-        },
-        {
-            description: <span>Scott McLeod, the "stand in bass player for the bass player" abruptly quits the band after their concert at Metropol in Pittsburgh, Pennsylvania, USA.  The band cancel the rest of their US tour dates and return to the UK after performing as four piece on the Late Show with David Letterman.</span>,
-            date: 'October 17',
-            source: { name: 'Official Oasis Site', title: 'Temporary Live Bassist Scott McLeod Leaves Oasis Tour', url: 'https://web.archive.org/web/19970605053852/http://www.oasisinet.com/byescott.htm' },
-            title: 'Scott McLeod Quits Oasis',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span><i>Don't Look Back In Anger</i> single cover photo session.  According to Brian Cannon, the cover is inspired by how George Martin covered Ringo Star's drumkit in flowers upon rejoining The Beatles after an argument.</span>,
-            date: 'October 17',
-            source: { name: '', title: '', url: '' },
-            title: 'Don\'t Look Back In Anger Cover Photo Session',
-            type: 'photo',
-            year: '1995'
-        },
-        {
-            description: <span>After Scott McLeod left the band abruptly during the tour, the band play <i>Morning Glory</i> as four piece with Bonehead on bass on the The Late Show with David Letterman.</span>,
-            date: 'October 19',
-            source: { name: 'Official Oasis Site', title: 'Videos - The Late Show with David Letterman', url: 'http://www.oasisinet.com/videos/whats-story-morning-glory-live-david-letterman-19-10-1995/' },
-            title: 'Oasis Perform Morning Glory on The Late Show with David Letterman',
-            type: 'noteworthy',
-            year: '1995'
-        },
-        {
-            description: <span><i>Wonderwall</i>, the third single from the album <i>(What's The Story) Morning Glory?</i>, is released on Creation Records.  The single includes the b-sides <i>Round Are Way</i>, <i>The Swamp Song</i> and <i>The Masterplan</i>.</span>,
-            date: 'October 30',
-            source: { name: 'Wikipedia', title: 'Wonderwall (song)', url: 'https://en.wikipedia.org/wiki/Wonderwall_(song)' },
-            title: 'Wonderwall Released',
-            type: 'releases',
-            year: '1995'
-        },
-        {
-            description: <span>The music video for <i>Wonderwall</i> is released.  The video is directed by Nigel Dick.</span>,
-            date: 'October 30',
-            source: { name: 'IMDB', title: 'Oasis: Wonderwall', url: 'https://www.imdb.com/title/tt5588998/' },
-            title: 'Wonderwall Music Video Released',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span>After recovering from nervous exhaustion (or "lazy-itis" according to Noel Gallagher), Guigsy re-joins the band after the departure of Scott McLeod to play in the first warm up gig at La Luna in Brussels, Belguim for the upcoming shows at Earls Court in London.</span>,
-            date: 'October 31',
-            source: { name: 'Dean Street Press', title: 'Getting High by Paolo Hewitt', url: 'http://deanstreetpress.co.uk/books/hewi02' },
-            title: 'Guigsy Returns - La Luna, Brussels, Belgium',
-            type: 'gigs',
-            year: '1995'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Definitely Maybe</i> album sales as 3x platinum.</span>,
-            date: 'November 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Definitely Maybe Sales Certified as 3x Platinum by BPI',
-            type: 'certifications',
-            year: '1995'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 3x platinum.</span>,
-            date: 'October 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 3x Platinum by BPI',
-            type: 'certifications',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis play two nights at Earls Court in London.  At the time, they are largets indoor gigs ever played in Europe.  Parts of the show appear in the official live video <i>...There and Then</i>.</span>,
-            date: 'November 4',
-            source: { name: 'Official Oasis Site', title: 'Gigography - Earls Court', url: 'http://www.oasisinet.com/#!/gigs/london-earls-court' },
-            title: 'Earls Court Exhibition Centre, London, UK - First Night',
-            type: 'gigs',
-            year: '1995'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 5x platinum.</span>,
-            date: 'December 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 5x Platinum by BPI',
-            type: 'certifications',
-            year: '1995'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Don't Look Back In Anger</i> at 1145 Arden Road in Pasadena, California, USA.  The video is directed by Nigel Dick in features Patrick Macnee who played John Steed on the 1960's television series <i>The Avengers</i>.</span>,
-            date: 'December 4',
-            source: { name: 'Nigel Dick Official Site', title: 'Productions 1995', url: 'http://www.nigeldick.com/films/productions-1995/' },
-            title: 'Don\'t Look Back In Anger Music Video Filmed',
-            type: 'videos',
-            year: '1995'
-        },
-        {
-            description: <span>The Recording Industry Association of America certifies <i>(What's The Story) Morning Glory?</i> album sales as gold.</span>,
-            date: 'January 4',
-            source: { name: 'Recording Industry Association of America', title: 'Gold & Platinum Awards', url: 'https://www.riaa.com/gold-platinum/?tab_active=default-award&ar=Oasis&ti=&lab=&genre=&format=&date_option=release&from=1993-01-01&to=1998-12-31&award=&type=&category=&adv=SEARCH#search_section' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as Gold by RIAA',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 6x platinum.</span>,
-            date: 'January 18',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 6x Platinum by BPI',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>The Recording Industry Association of America certifies <i>(What's The Story) Morning Glory?</i> album sales as platinum.</span>,
-            date: 'January 24',
-            source: { name: 'Recording Industry Association of America', title: 'Gold & Platinum Awards', url: 'https://www.riaa.com/gold-platinum/?tab_active=default-award&ar=Oasis&ti=&lab=&genre=&format=&date_option=release&from=1993-01-01&to=1998-12-31&award=&type=&category=&adv=SEARCH#search_section' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as Platinum by RIAA',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 7x platinum.</span>,
-            date: 'February 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 7x Platinum by BPI',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Champagne Supernova</i> at Ealing Studios in London, UK.  The shoot spans two days and is directed by Nigel Dick.</span>,
-            date: 'February 16',
-            source: { name: 'Nigel Dick Official Site', title: 'Productions 1996', url: 'http://www.nigeldick.com/films/productions-1996/' },
-            title: 'Champagne Supernova Music Video Filmed',
-            type: 'videos',
-            year: '1996'
-        },
-        {
-            description: <span>Oasis beat out rival Blur for the British Group, British Album of the Year and British Video of the Year BRIT awards.  During their acceptance speech for British Group of the Year, Noel and Liam performed a taunting version of Blur's <i>Parklife</i>.  The award ceremony is broadcast by ITV a few days later on February 22.</span>,
-            date: 'February 19',
-            source: { name: 'The Independent', title: 'Blur and Oasis in clash for Brits', url: 'https://www.independent.co.uk/news/blur-and-oasis-in-clash-for-brits-1323071.html' },
-            title: 'Oasis Dominate the BRIT Awards',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span><i>Don't Look Back In Anger</i>, the fourth single from the album <i>(What's The Story) Morning Glory?</i>, is released on Creation Records.  The single includes the b-sides <i>Step Out</i>, <i>Underneath the Sky</i> and a cover of <i>Cum On Feel The Noize</i> originally by Slade.</span>,
-            date: 'February 19',
-            source: { name: 'Wikipedia', title: 'Don\'t Look Back In Anger', url: 'https://en.wikipedia.org/wiki/Don%27t_Look_Back_in_Anger' },
-            title: 'Don\'t Look Back In Anger Released',
-            type: 'releases',
-            year: '1996'
-        },
-        {
-            description: <span>The music video for <i>Don't Look Back In Anger</i> is released.  The video is directed by Nigel Dick.</span>,
-            date: 'February 19',
-            source: { name: 'IMDB', title: 'Oasis: Don\'t Look Back In Anger', url: 'https://www.imdb.com/title/tt5675280/' },
-            title: 'Don\'t Look Back In Anger Music Video Released',
-            type: 'videos',
-            year: '1996'
-        },
-        {
-            description: <span>As part of an effort to find a venue big enough for Oasis to play during the summer due to a fallow year for the Glastonbury Festival, Noel Gallagher visits Knebworth House in Stevenage, UK.  Upon visiting, Noel is convinced that it would be the site of their biggest concert of the summer.</span>,
-            date: 'March',
-            source: { name: 'GQ UK', title: 'Oasis: Supersonic - Noel and Liam Gallagher speak exclusively to GQ', url: 'https://www.gq-magazine.co.uk/article/oasis-supersonic-noel-liam-gallagher' },
-            title: 'Noel Gallagher Visits Knebworth House',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Definitely Maybe</i> album sales as 4x platinum.</span>,
-            date: 'March 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Definitely Maybe Sales Certified as 4x Platinum by BPI',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 8x platinum.</span>,
-            date: 'March 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 8x Platinum by BPI',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>The Recording Industry Association of America certifies <i>(What's The Story) Morning Glory?</i> album sales as 2x platinum.</span>,
-            date: 'March 20',
-            source: { name: 'Recording Industry Association of America', title: 'Gold & Platinum Awards', url: 'https://www.riaa.com/gold-platinum/?tab_active=default-award&ar=Oasis&ti=&lab=&genre=&format=&date_option=release&from=1993-01-01&to=1998-12-31&award=&type=&category=&adv=SEARCH#search_section' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 2x Platinum by RIAA',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>British tabloid <i>News of the World</i> pays Noel and Liam Gallagher's estranged father, Tommy, to drop in to the Dublin Westbury Hotel in Dublin Ireland where the band is celebrating after their second night of shows at The Point in Dublin.</span>,
-            date: 'March 23',
-            source: { name: 'The Daily Edge', title: 'Those legendary 1996 Oasis Dublin Point gigs', url: 'https://www.dailyedge.ie/oasis-supersonic-dublin-point-3007282-Oct2016/' },
-            title: 'News of the World Pays Tommy Gallagher to "Reconcile" With His Sons',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>Oasis play the first to two sell-out nights at the former home of Manchester City Football Club.  It represents a homecoming for the band.  Tickets sold out the 40,000 seat stadium in under 3 hours when they went on sale in February, leading to the addition of a second show.  Parts of the show appear in the official live video <i>...There and Then</i>.</span>,
-            date: 'April 27',
-            source: { name: 'Manchester Evening News', title: 'Oasis at Maine Road 1996: seven things Oasis were absolutely nailing that year', url: 'https://www.manchestereveningnews.co.uk/whats-on/music-nightlife-news/oasis-maine-road-1996-moments-11245625' },
-            title: 'Maine Road, Manchester, UK - First Night',
-            type: 'gigs',
-            year: '1996'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 9x platinum.</span>,
-            date: 'May 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 9x Platinum by BPI',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>Oasis appear on the cover of Rolling Stone Magazine for the first time.  The <a href="https://www.rollingstone.com/music/music-news/oasis-ruling-asses-243843/" target="_blank" rel="noopener noreferrer">accompanying story</a> is not so flattering, focusing more on the band's extracurricular activities.</span>,
-            date: 'May 2',
-            source: { name: 'Rolling Stone', title: 'Oasis: Ruling Asses', url: 'https://www.rollingstone.com/music/music-news/oasis-ruling-asses-243843/' },
-            title: 'Oasis Appear on Cover of Rolling Stone Magazine',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>Oasis appear on Chris Evans' Radio 1 Breakfast Show and announce that they will be playing two nights at Knebworth in August.</span>,
-            date: 'May 7',
-            source: { name: 'GQ UK', title: 'Oasis: Supersonic - Noel and Liam Gallagher speak exclusively to GQ', url: 'https://www.gq-magazine.co.uk/article/oasis-supersonic-noel-liam-gallagher' },
-            title: 'Oasis Announce Knebworth Concerts',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>Tickets for Oasis' gig at Knebworth House, Stevenage, UK go on sale at 9AM.  2.6 million people (or 5% of the British population) apply for tickets, leading to the addition of a second night.  Based on the demand, Oasis could have sold out the venue for two weeks.</span>,
-            date: 'May 11',
-            source: { name: 'RadioX', title: '10 Facts About Oasis At Knebworth, 1996', url: 'https://www.radiox.co.uk/artists/oasis/10-facts-about-oasis-at-knebworth-1996/' },
-            title: 'Knebworth Tickets Go On Sale',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>Noel Gallagher begins his vacation on the Caribbean island of Mustique.  While staying at Mick Jagger's residence with girlfriend Meg Matthews, model Kate Moss and Johnny Depp, Noel records demos for <i>Be Here Now</i>.  After two weeks, Owen Morris arrives in early June to help Noel complete the demos.  These demos are included on the 2016 Chasing the Sun re-issue of <i>Be Here Now</i>.</span>,
-            date: 'May 26',
-            source: { name: 'Dean Street Press', title: 'Getting High by Paolo Hewitt', url: 'http://deanstreetpress.co.uk/books/hewi02' },
-            title: 'Mustique Demos Recorded',
-            type: 'recordings',
-            year: '1996'
-        },
-        {
-            description: <span>The Recording Industry Association of America certifies <i>(What's The Story) Morning Glory?</i> album sales as 3x platinum.</span>,
-            date: 'June 20',
-            source: { name: 'Recording Industry Association of America', title: 'Gold & Platinum Awards', url: 'https://www.riaa.com/gold-platinum/?tab_active=default-award&ar=Oasis&ti=&lab=&genre=&format=&date_option=release&from=1993-01-01&to=1998-12-31&award=&type=&category=&adv=SEARCH#search_section' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 3x Platinum by RIAA',
-            type: 'certifications',
-            year: '1996'
-        },
-        {	description: <span>Noel Gallagher performs <i>This Guy's In Love With You</i> with Burt Bacharach at Royal Festival Hall, London, UK.  It is the only time Noel has performed live without playing an instrument.</span>,
-            date: 'June 28',
-            source: { name: 'Oasis Interviews', title: 'Noel Gallagher Interview August 11, 1996 - BBC Radio 1', url: 'http://oasisinterviews.blogspot.com/1996/08/' },
-            title: 'Noel Gallagher Performs with Burt Bacharach at London\'s Royal Festival Hall',
-            type: 'gigs',
-            year: '1996'
-        },
-        {
-            description: <span>The band and their management officially file to trademark the word "Oasis" in association with a band/music in the UK.</span>,
-            date: 'July 26',
-            source: { name: 'Intellectual Property Office', title: 'Search for a trade mark', url: 'https://trademarks.ipo.gov.uk/ipo-tmcase/page/Results/4/EU000347666' },
-            title: 'Band Files to Trademark term \'Oasis\' in the UK',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 10x platinum.</span>,
-            date: 'August 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 10x Platinum by BPI',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>A lorry driver dies during set up for the upcoming concert when he is crushed between his vehicle and a forklift.  Oasis were reportedly practicing on stage at the time of the incident.  Despite the incident, Oasis go on with the two sold out shows, performing to 80,000 people over the two days.</span>,
-            date: 'August 2',
-            source: { name: 'The Herald', title: 'Oasis concert driver killed', url: 'https://www.heraldscotland.com/news/12114715.oasis-concert-driver-killed/' },
-            title: 'Worker Killed While Setting Up for Loch Lomond Concert',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>Oasis play the first to two sell-out nights at Knebworth House, Hertfordshire, UK.  Oasis play to 125,000 people each night, with a guest list of 7,000.  They debuted two new songs which would later be included on their third studio album, <i>Be Here Now</i>:  <i>It's Gettin' Better (Man!!)</i> and <i>My Big Mouth</i>.  Support acts included The Charlatans, Manic Street Preachers, The Prodigy, Kula Shaker, Cast and Dreadzone.</span>,
-            date: 'August 10',
-            source: { name: 'BBC Music', title: '20 years on: 10 staggering facts about Oasis at Knebworth', url: 'https://www.bbc.co.uk/music/articles/2868f61d-ea81-46b4-8c19-9fc4974ec359' },
-            title: 'Knebworth House, Hertfordshire, UK - First Night',
-            type: 'gigs',
-            year: '1996'
-        },
-        {
-            description: <span>Without lead singer Liam Gallagher, Oasis performs at Royal Festival Hall in London, UK for MTV's <i>Unplugged</i> series with Noel Gallagher performing lead vocals.  Liam bailed out at the last minute claiming that he had a sore throat.  That did not stop him from heckling brother Noel from the balcony during the taping.</span>,
-            date: 'August 23',
-            source: { name: 'BlastEcho', title: 'The Night Oasis Unplugged Without Liam Gallagher', url: 'http://blastecho.com/the-night-oasis-unplugged-without-liam-gallagher/' },
-            title: 'Oasis Unplugged - Royal Festival Hall, London, UK',
-            type: 'gigs',
-            year: '1996'
-        },
-        {
-            description: <span>Oasis begin a US tour at the Rosemont Horizon in Chicago, Illinois, USA.  The show is noteworthy because Liam Gallagher did not accompany the band to the US, instead staying in the UK to look for houses with his girlfriend Patsy Kensit.  Noel Gallagher performs vocals.  Liam rejoins the band for their next show on August 30 in Detroit, Michigan, USA.</span>,
-            date: 'August 27',
-            source: { name: 'The Independent', title: 'Oasis tour of US hit as singer pulls out', url: 'https://www.independent.co.uk/news/oasis-tour-of-us-hit-as-singer-pulls-out-1311642.html' },
-            title: 'Oasis Begin US Tour Without Liam',
-            type: 'gigs',
-            year: '1996'
-        },
-        {
-            description: <span>During a performance of <i>Champagne Supernova</i> at the MTV Music Awards in New York, New York, USA, Liam Gallagher infamously swears at the audience, spits on stage and hurls an open beer can into the crowd.</span>,
-            date: 'September 4',
-            source: { name: 'MTV', title: 'Video Music Awards 96', url: 'http://www.mtv.com/vma/1996' },
-            title: 'Oasis Antics at MTV Music Awards in New York',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>After a fight between Liam and Noel and issues with a show in South Carolina, USA, Noel Gallagher leaves the tour and flies home to the UK.  All subsequent dates in the US and Australia are cancelled.</span>,
-            date: 'September 12',
-            source: { name: 'Official Oasis Site', title: 'Official Statement - US Tour Cancellation', url: 'https://web.archive.org/web/19970605053809/http://www.oasisinet.com/noasis.htm' },
-            title: 'Oasis Cancels Remaining US Tour Dates',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>Despite media speculation in the wake of their cancelled US tour, Oasis announce that they will continue to exist as a band.</span>,
-            date: 'September 13',
-            source: { name: 'Official Oasis Site', title: 'Official Statement - US Tour Cancellation', url: 'https://web.archive.org/web/19970605053809/http://www.oasisinet.com/noasis.htm' },
-            title: 'Oasis Announce That They Have Not Split Up',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>After a disastrous US tour, Oasis announce that they are beginning work on their third studio album.</span>,
-            date: 'September 17',
-            source: { name: 'Official Oasis Site', title: 'Official Statement - New Album', url: 'https://web.archive.org/web/19971211000255/http://www.oasisinet.com/prokasis.htm' },
-            title: 'Oasis Announce Work on Third Studio Album',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Definitely Maybe</i> album sales as 5x platinum.</span>,
-            date: 'October 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Definitely Maybe Sales Certified as 5x Platinum by BPI',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 11x platinum.</span>,
-            date: 'October 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 11x Platinum by BPI',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>Fresh off the disastrous US tour, Oasis book <a href="https://www.abbeyroad.com/" target="_blank" rel="noopener noreferrer">Abbey Road Studios</a> in London to being recording their third studio album.</span>,
-            date: 'October 7',
-            source: { name: 'Q Magazine', title: 'The Oasis Diaries September, 1997', url: 'https://www.rocksbackpages.com/Library/Article/piece-of-piss-the-oasis-diaries' },
-            title: 'Oasis Begin Recording Be Here Now at Abbey Road Studios',
-            type: 'recordings',
-            year: '1996'
-        },
-        {
-            description: <span>Officially released live video mixing interviews of the band with performances from Earls Court, London, UK (4/5 November, 1995) and Maine Road, Manchester, UK (April 28, 1996).</span>,
-            date: 'October 14',
-            source: { name: 'Wikipedia', title: '...There and Then', url: 'https://en.wikipedia.org/wiki/...There_and_Then' },
-            title: '...There and Then Released',
-            type: 'releases',
-            year: '1996'
-        },
-        {
-            description: <span>A CD box set containing the four singles from <i>Definitely Maybe</i> plus an interview disc is released in a collectible box inspired by a Benson & Hedges cigarette box.  The box bares the slogan: "Warning: Rock 'N' Roll Can Seriously Damage Your Health."</span>,
-            date: 'November 5',
-            source: { name: 'Wikipedia', title: 'Definitely Maybe', url: 'https://en.wikipedia.org/wiki/Definitely_Maybe#Singles_box_set' },
-            title: 'Definitely Maybe: Singles Released',
-            type: 'releases',
-            year: '1996'
-        },
-        {
-            description: <span>The Recording Industry Association of America certifies <i>(What's The Story) Morning Glory?</i> album sales as 4x platinum.</span>,
-            date: 'November 8',
-            source: { name: 'Recording Industry Association of America', title: 'Gold & Platinum Awards', url: 'https://www.riaa.com/gold-platinum/?tab_active=default-award&ar=Oasis&ti=&lab=&genre=&format=&date_option=release&from=1993-01-01&to=1998-12-31&award=&type=&category=&adv=SEARCH#search_section' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 4x Platinum by RIAA',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>During the recording of Oasis' third studio album, <i>Be Here Now</i>, at <a href="https://www.abbeyroad.com/" target="_blank" rel="noopener noreferrer">Abbey Road Studios</a> in London, UK, Liam is arrested on suspicion of possession of a controlled substance.  Charges would later be dropped when the substance turned out to be fake.</span>,
-            date: 'November 9',
-            source: { name: 'The Independent', title: 'Anger as Oasis star cautioned over drug', url: 'https://www.independent.co.uk/news/anger-as-oasis-star-cautioned-over-drug-1282122.html' },
-            title: 'Liam Gallagher Arrested in London\'s West End',
-            type: 'noteworthy',
-            year: '1996'
-        },
-        {
-            description: <span>A CD box set containing the four singles from <i>(What's The Story) Morning Glory?</i> plus an interview disc is released in a collectible box inspired by a Benson & Hedges cigarette box.  The box bares the slogan: "Warning: Rock 'N' Roll Can Seriously Damage Your Health."</span>,
-            date: 'November 10',
-            source: { name: 'Wikipedia', title: '(What\'s The Story) Morning Glory?', url: 'https://en.wikipedia.org/wiki/(What%27s_the_Story)_Morning_Glory%3F#Singles_box_set' },
-            title: '(What\'s The Story) Morning Glory: Singles Released',
-            type: 'releases',
-            year: '1996'
-        },
-        {
-            description: <span>Issues with studio management, combined with staff members leaking recordings to the press, force Oasis to leave <a href="https://www.abbeyroad.com/" target="_blank" rel="noopener noreferrer">Abbey Road Studios</a>.  They relocate to <a href="https://www.ridgefarm.com/history/" target="_blank" rel="noopener noreferrer">Ridge Farm Studios</a> in Surrey, UK.</span>,
-            date: 'November 11',
-            source: { name: 'Rock\'s Back Pages', title: 'Piece of Piss: The Oasis Diaries', url: 'https://www.rocksbackpages.com/Library/Article/piece-of-piss-the-oasis-diaries' },
-            title: 'Oasis Kicked Out of Abbey Road Studios',
-            type: 'recordings',
-            year: '1996'
-        },
-        {
-            description: <span>After being forced to abandon the recording sessions at <a href="https://www.abbeyroad.com/" target="_blank" rel="noopener noreferrer">Abbey Road Studios</a> in London, Oasis relocate to <a href="https://www.ridgefarm.com/history/" target="_blank" rel="noopener noreferrer">Ridge Farm Studios</a> in Surrey, UK to continue the recording of <i>Be Here Now</i>.</span>,
-            date: 'November 18',
-            source: { name: 'Be Here Now', title: 'Fanclub Boxset Booklet', url: 'http://i.imgur.com/5NnIDan.jpg' },
-            title: 'Oasis Resume Recording Be Here Now at Ridge Farm Studios',
-            type: 'recordings',
-            year: '1996'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 12x platinum.</span>,
-            date: 'December 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 12x Platinum by BPI',
-            type: 'certifications',
-            year: '1996'
-        },
-        {
-            description: <span>After a productive session at <a href="https://www.ridgefarm.com/history/" target="_blank" rel="noopener noreferrer">Ridge Farm Studios</a> in Surrey, UK, Oasis reconvene after the holidays to finish recording of <i>Be Here Now</i> at <a href="http://www.airstudios.com/" target="_blank" rel="noopener noreferrer">Air Recording Studios</a> in London, UK.</span>,
-            date: 'January 10',
-            source: { name: 'Be Here Now', title: 'Fanclub Boxset Booklet', url: 'http://i.imgur.com/5NnIDan.jpg' },
-            title: 'Oasis Continue Recording Be Here Now at Air Studios',
-            type: 'recordings',
-            year: '1997'
-        },
-        {
-            description: <span>Speaking to BBC Radio 1 about the ouster of Brian Harvey from East 17, Noel utters his famous quote that taking drugs is "like getting up and having a cup of tea in the morning."  He is immediately condemned by politicians.</span>,
-            date: 'January 30',
-            source: { name: 'The Independent', title: 'Drugs are like a cup of tea, says Noel', url: 'https://www.independent.co.uk/news/drugs-are-like-a-cup-of-tea-says-noel-1285833.html' },
-            title: 'Noel Gallagher Sparks Outrage over Drug Comments',
-            type: 'noteworthy',
-            year: '1997'
-        },
-        {
-            description: <span>Oasis begin the final stretch of recording for <i>Be Here Now</i> at <a href="http://www.airstudios.com/" target="_blank" rel="noopener noreferrer">Air Recording Studios</a> in London, UK.  The mood is somber, as during the session Paul "Bonehead" Arthur's mother passes away.</span>,
-            date: 'February 1',
-            source: { name: 'Rock\'s Back Pages', title: 'Piece of Piss: The Oasis Diaries', url: 'https://www.rocksbackpages.com/Library/Article/piece-of-piss-the-oasis-diaries' },
-            title: 'Oasis Continue Recording Be Here Now at Air Studios',
-            type: 'recordings',
-            year: '1997'
-        },
-        {
-            description: <span>Liam Gallagher and Patsy Kenset call off their first attempt to get married after the British press discovery the location and details ahead of time.</span>,
-            date: 'February 14',
-            source: { name: 'Oasis FAQ', title: 'Personal Details', url: 'http://www.oocities.org/sunsetstrip/underground/3284/persdets.html?201920' },
-            title: 'Wedding Between Liam Gallagher and Patsy Kinset Called Off Due to Media Leak',
-            type: 'noteworthy',
-            year: '1997'
-        },
-        {
-            description: <span>Oasis begin the final recording session for <i>Be Here Now</i> at Master Rock Studios in Kilburn, London, UK.</span>,
-            date: 'February 26',
-            source: { name: 'Be Here Now', title: 'Fanclub Boxset Booklet', url: 'http://i.imgur.com/5NnIDan.jpg' },
-            title: 'Final Recording Session for Be Here Now Begins at Master Rock Studios',
-            type: 'recordings',
-            year: '1997'
-        },
-        {
-            description: <span><i>D'You Know What I Mean?</i> single cover photo session in Wigan, Greater Manchester, UK.</span>,
-            date: 'March 19',
-            source: { name: 'Sonic Editions', title: 'D\'You Know What I Mean?', url: 'https://www.soniceditions.com/image/oasis-do-you-know' },
-            title: 'D\'You Know What I Mean? Cover Photo Session',
-            type: 'photo',
-            year: '1997'
-        },
-        {
-            description: <span>Owen Morris and Noel Gallagher begin mixing songs for <i>Be Here Now</i> at <a href="https://milocostudios.com/studios/orinoco-complex/" target="_blank" rel="noopener noreferrer">Orinoco Studios</a> in Bermondesey, London, UK.</span>,
-            date: 'March 20',
-            source: { name: 'Be Here Now', title: 'Fanclub Boxset Booklet', url: 'http://i.imgur.com/5NnIDan.jpg' },
-            title: 'Mixing of Be Here Now at Orinoco Studios',
-            type: 'recordings',
-            year: '1997'
-        },
-        {
-            description: <span>Paul "Guigsy" McGuigan marries girlfriend Ruth Tolhurst in St. Lucia in the Caribbean.  The couple met on a flight during Oasis' first tour of Japan in September 1994.</span>,
-            date: 'April 4',
-            source: { name: 'The Free Library', title: 'GUIGSY\'S A SPLICE BOY; Noel\'s pal weds in secret', url: 'https://www.thefreelibrary.com/GUIGSY%27S+A+SPLICE+BOY%3B+Noel%27s+pal+weds+in+secret.-a061100473' },
-            title: 'Paul "Guigsy" McGuigan Marries Ruth Tolhurst',
-            type: 'noteworthy',
-            year: '1997'
-        },
-        {
-            description: <span>Liam Gallagher marries girlfriend Patsy Kensit in the Marylebone Registery Office in London, UK.  It is the same place where Paul and Linda McCarthy got married in 1969.</span>,
-            date: 'April 7',
-            source: { name: 'Official Oasis Site', title: 'Liam Marries Fiance Patsy Kensit', url: 'https://web.archive.org/web/19970605051757/http://www.oasisinet.com/mrsliam.htm' },
-            title: 'Liam Gallagher and Patsy Kensit Wed',
-            type: 'noteworthy',
-            year: '1997'
-        },
-        {
-            description: <span>Oasis announce that their third studio album will be released in August.  The title is confirmed as <i>Be Here Now</i>.  The band also announces that 15 tracks were recorded, 11 of which will be included on the album.</span>,
-            date: 'May 15',
-            source: { name: 'Official Oasis Site', title: 'New Oasis LP BE HERE NOW', url: 'https://web.archive.org/web/19970605051723/http://www.oasisinet.com/beherepr.htm' },
-            title: 'Oasis Announce Release Details for Be Here Now',
-            type: 'noteworthy',
-            year: '1997'
-        },
-        {
-            description: <span>Under secrecy, Oasis shoot the cover for <i>Be Here Now</i> at Stocks House in Hertfordshire, UK with Michael Spencer Jones.  Originally, the cover is to be shot at night time, but due to issues with a generator, a backup shot from earlier in the day is selected.</span>,
-            date: 'May 16',
-            source: { name: 'Sonic Editions', title: 'Be Here Now', url: 'https://www.soniceditions.com/image/oasis-be-here-now' },
-            title: 'Be Here Now Album Cover Photo Session',
-            type: 'photo',
-            year: '1997'
-        },
-        {
-            description: <span>Noel Gallagher marries girlfriend Meg Matthews at the Little Church of the West in Las Vegas, Nevada, USA.</span>,
-            date: 'June 6',
-            source: { name: 'Official Oasis Site', title: 'Mr & Mrs Noel Gallagher', url: 'https://web.archive.org/web/19991114154438/http://www.oasisinet.com/mrsnoel.htm' },
-            title: 'Noel Gallagher and Meg Matthews Wed',
-            type: 'noteworthy',
-            year: '1997'
-        },
-        {
-            description: <span>As a favour the the Beastie Boys, Noel Gallagher plays a short, electric set of five songs at the "Free Tibet" concert on Randall's Island Park in New York, New York, USA.  The set includes the first live performance of <i>Helter Skelter</i> by Noel.  The concert would cause trouble for Oasis in 2009 when the band attempted to play in China for the first time.</span>,
-            date: 'June 7',
-            source: { name: 'Wikipedia', title: 'Tibetan Freedom Concert', url: 'https://en.wikipedia.org/wiki/Tibetan_Freedom_Concert#New_York_City' },
-            title: 'Noel Gallagher Performs at the "Free Tibet" Concert in New York',
-            type: 'gigs',
-            year: '1997'
-        },
-        {
-            description: <span>Oasis begins their world tour in support of their yet to be released third studio album <i>Be Here Now</i> at the KROQ Weenie Roast in support of U2 in Irvine, California, USA.</span>,
-            date: 'June 14',
-            source: { name: 'Wikipedia', title: 'Be Here Now Tour', url: 'https://en.wikipedia.org/wiki/Be_Here_Now_Tour' },
-            title: 'Be Here Now World Tour Begins',
-            type: 'gigs',
-            year: '1997'
-        },
-        {
-            description: <span>Oasis film the music video for <i>D'You Know What I Mean?</i> at the former Gas Works in Becton, East London, UK.  The video is directed by Dom & Nick Goffey.</span>,
-            date: 'June 24',
-            source: { name: 'Live4Ever Forum', title: 'Oasis\' Videos', url: 'http://live4ever.proboards.com/post/393777/thread' },
-            title: 'D\'You Know What I Mean? Music Video Filmed',
-            type: 'videos',
-            year: '1997'
-        },
-        {
-            description: <span><i>D'you Know What I Mean?</i>, the first single from the album <i>Be Here Now</i>, is released on Creation Records.  The single includes the b-sides <i>Stay Young</i>, <i>Angel Child (demo)</i> and a cover of <i>Heroes</i> originally by David Bowie.</span>,
-            date: 'July 7',
-            source: { name: 'Wikipedia', title: 'D\'You Know What I Mean?', url: 'https://en.wikipedia.org/wiki/D%27You_Know_What_I_Mean%3F' },
-            title: 'D\'You Know What I Mean? Released',
-            type: 'releases',
-            year: '1997'
-        },
-        {
-            description: <span>The music video for <i>D'You Know What I Mean?</i> is released.  The video is directed by "Dom & Nic" (Nick Goffey and Dominic Hawley).</span>,
-            date: 'July 7',
-            source: { name: 'IMDB', title: 'Oasis: D\'You Know What I Mean?', url: 'https://www.imdb.com/title/tt5675786/' },
-            title: 'D\'You Know What I Mean? Music Video Released',
-            type: 'videos',
-            year: '1997'
-        },
-        {
-            description: <span>Noel Gallagher, along with wife Meg Matthews and Creation Records boss Alan McGee, attend an event for new elected Prime Mister Tony Blair at the Prime Minster's residence at 10 Downing Street in London, UK.</span>,
-            date: 'July 30',
-            source: { name: 'Information Britain', title: 'Cool Britannia Party at 10 Downing Street', url: 'http://www.information-britain.co.uk/famdates.php?id=1121' },
-            title: 'Noel Gallagher Attends 10 Downing Street Event',
-            type: 'noteworthy',
-            year: '1997'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Be Here Now</i> album sales as 5x platinum.</span>,
-            date: 'August 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Be Here Now Sales Certified as 5x Platinum by BPI',
-            type: 'certifications',
-            year: '1997'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Don't Go Away</i> in Chertsey and London, UK over two days.  The video is directed by Nigel Dick.</span>,
-            date: 'August 11',
-            source: { name: 'Nigel Dick Official Site', title: 'Productions 1997', url: 'http://www.nigeldick.com/films/productions-1997/' },
-            title: 'Don\'t Go Away Music Video Filmed',
-            type: 'videos',
-            year: '1997'
-        },
-        {
-            description: <span>Alan White and Liz Atkins marry at the Studley Priory Hotel in Oxfordshire, UK.  The couple met during the filming of the promotional video for <i>Don't Look Back In Anger</i> in Bel Air, California, USA.</span>,
-            date: 'August 13',
-            source: { name: 'Drummers Zone', title: 'Alan White Biography', url: 'https://drummerszone.com/artists/profile/10784/alan-white#biography' },
-            title: 'Alan White and Liz Atkins Wed',
-            type: 'noteworthy',
-            year: '1997'
-        },
-        {
-            description: <span>Oasis' third studio album <i>Be Here Now</i> is released by Creation Records.  The album would go on to set the record at the time for selling the most units on the first day of release.</span>,
-            date: 'August 21',
-            source: { name: 'Wikipedia', title: 'Be Here Now (album)', url: 'https://en.wikipedia.org/wiki/Be_Here_Now_(album)' },
-            title: 'Be Here Now Released',
-            type: 'releases',
-            year: '1997'
-        },
-        {
-            description: <span><i>Stand By Me</i>, the second single from the album <i>Be Here Now</i>, is released on Creation Records.  The single includes the b-sides <i>(I Got) The Fever</i>, <i>My Sister Lover</i> and <i>Going Nowhere</i>.</span>,
-            date: 'September 22',
-            source: { name: 'Wikipedia', title: 'Stand By Me (Oasis song)', url: 'https://en.wikipedia.org/wiki/Stand_by_Me_(Oasis_song)' },
-            title: 'Stand By Me Released',
-            type: 'releases',
-            year: '1997'
-        },
-        {
-            description: <span>The music video for <i>Stand By Me</i> is released.  The video is directed by David Mould.</span>,
-            date: 'September 22',
-            source: { name: 'IMDB', title: 'Oasis: Stand By Me', url: 'https://www.imdb.com/title/tt7078344/' },
-            title: 'Stand By Me Music Video Released',
-            type: 'videos',
-            year: '1997'
-        },
-        {
-            description: <span>The music video for <i>Don't Go Away</i> is released.  The video is directed by Nigel Dick.</span>,
-            date: 'October',
-            source: { name: 'IMDB', title: 'Oasis: Don\'t Go Away', url: 'https://www.imdb.com/title/tt5675468/' },
-            title: 'Don\'t Go Away Music Video Released',
-            type: 'videos',
-            year: '1997'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 13x platinum.</span>,
-            date: 'October 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 13x Platinum by BPI',
-            type: 'certifications',
-            year: '1997'
-        },
-        {
-            description: <span>The Recording Industry Association of America certifies <i>Be Here Now</i> album sales as platinum.</span>,
-            date: 'November 3',
-            source: { name: 'Recording Industry Association of America', title: 'Gold & Platinum Awards', url: 'https://www.riaa.com/gold-platinum/?tab_active=default-award&ar=Oasis&ti=&lab=&genre=&format=&date_option=release&from=1993-01-01&to=1998-12-31&award=&type=&category=&adv=SEARCH#search_section' },
-            title: 'Be Here Now Sales Certified as Platinum by RIAA',
-            type: 'certifications',
-            year: '1997'
-        },
-        {
-            description: <span>Oasis abruptly end a concert at the S.E.C.C. in Glasgow, Scotland, UK after Paul "Bonehead" Arthurs is struck by a bottle thrown at the stage.  The band played 10 songs before walking off stage and ending the gig.</span>,
-            date: 'December 7',
-            source: { name: 'The Herald', title: 'Oasi storm off in Glasgow after bottle thrown', url: 'https://www.heraldscotland.com/news/12349347.oasis-storm-off-in-glasgow-after-bottle-thrown/' },
-            title: 'Gig at the S.E.C.C. in Glasgow Ends After Bottle Thrown on Stage',
-            type: 'gigs',
-            year: '1997'
-        },
-        {
-            description: <span><i>All Around the World</i>, the third and final single from the album <i>Be Here Now</i>, is released on Creation Records.  It still holds the record for the longest song to reach number on in the UK.  The single includes the b-sides <i>The Fame</i>, <i>Flashbax</i> and a cover of <i>Stream Fighting Man</i> originally by the Rolling Stones.</span>,
-            date: 'January 12',
-            source: { name: 'Wikipedia', title: 'All Around The World (Oasis song)', url: 'https://en.wikipedia.org/wiki/All_Around_the_World_(Oasis_song)' },
-            title: 'All Around the World Released',
-            type: 'releases',
-            year: '1998'
-        },
-        {
-            description: <span>The music video for <i>All Around the World</i> is released.  The video is directed by Jonathan Dayton and Valerie Faris.</span>,
-            date: 'January 12',
-            source: { name: 'IMDB', title: 'Oasis: All Around the World', url: 'https://www.imdb.com/title/tt1616136/' },
-            title: 'All Around the World Music Video Released',
-            type: 'videos',
-            year: '1998'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Be Here Now</i> album sales as 6x platinum.</span>,
-            date: 'January 30',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Be Here Now Sales Certified as 6x Platinum by BPI',
-            type: 'certifications',
-            year: '1998'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Definitely Maybe</i> album sales as 6x platinum.</span>,
-            date: 'January 30',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Definitely Maybe Sales Certified as 6x Platinum by BPI',
-            type: 'certifications',
-            year: '1998'
-        },
-        {
-            description: <span>Liam Gallagher is banned for life by airline Cathay Pacific after allegedly throwing a scone at a flight attendant on a flight from Hong Kong to Perth, Australia.</span>,
-            date: 'February 23',
-            source: { name: 'MTV', title: 'Oasis\' Liam Banned From Airline ', url: 'http://www.mtv.com/news/1432831/oasis-liam-banned-from-airline/' },
-            title: 'Liam Banned by Cathay Pacific Airlines',
-            type: 'noteworthy',
-            year: '1998'
-        },
-        {
-            description: <span>Liam Gallagher is arrested outside of a Brisbane, Australia hotel after he is accused of head-butting a fan, Ben Jones, who attempted to take Liam's picture.</span>,
-            date: 'March 5',
-            source: { name: 'BBC World News', title: 'Oasis singer banned by airline', url: 'http://news.bbc.co.uk/2/hi/asia-pacific/64844.stm' },
-            title: 'Liam Arrested for Headbutting Fan in Australia',
-            type: 'noteworthy',
-            year: '1998'
-        },
-        {
-            description: <span>Unknown at the time, Oasis play their last live gig with original founding members Paul "Bonehead" Arthurs and Paul "Guigsy" McGuigan at the Palacio de los Deportes in Mexico City, Mexico.  Prior to their next tour to support their new album, both members depart the band.</span>,
-            date: 'March 28',
-            source: { name: 'Wikipedia', title: 'Be Here Now Tour', url: 'https://en.wikipedia.org/wiki/Be_Here_Now_Tour' },
-            title: 'Last Gig for Bonehead and Guigsy',
-            type: 'gigs',
-            year: '1998'
-        },
-        {
-            description: <span>During halftime of the World Cup match between Germany and Greece, Noel Gallagher decides that he is done with drugs.  Over the next few months, he would sell his Belsize Park home in London, spend a month vacationing in Thailand and move to the English country side in order to kick his drug habit.</span>,
-            date: 'June 5',
-            source: { name: 'The Guardian', title: 'Noel Gallagher Interview', url: 'https://www.theguardian.com/observer/omm/story/0,,1507223,00.html'},
-            title: 'Noel Gallagher Gives Up Drugs',
-            type: 'noteworthy',
-            year: '1998'
-        },
-        {
-            description: <span>Noel Gallagher begins recording demos of the songs that will feature on Oasis' upcoming fourth studio album at Wheeler End Studios in Buckinghamshire, UK with engineer Stuart Epps.</span>,
-            date: 'October 5',
-            source: { name: 'Official Oasis Site', title: 'Official News Updates', url: 'https://web.archive.org/web/19981201072647/http://www.oasisinet.com/news.htm' },
-            title: 'Demo Sessions for Standing on the Shoulder of Giants',
-            type: 'recordings',
-            year: '1998'
-        },
-        {
-            description: <span><i>The Masterplan</i>, a compilation album of single b-sides from Oasis' first three studio albums, is released.  It is the last Oasis album released by Creation records.</span>,
-            date: 'November 3',
-            source: { name: 'Wikipedia', title: 'The Masterplan (album)', url: 'https://en.wikipedia.org/wiki/The_Masterplan_(album)' },
-            title: 'The Masterplan Released',
-            type: 'releases',
-            year: '1998'
-        },
-        {
-            description: <span>Liam Gallagher is arrested by Metropolitan Police in London, UK after a fight with a photographer resulted in a broken camera.  Liam is brought to Kentish Town police station for questioning and released on bail.</span>,
-            date: 'November 4',
-            source: { name: 'MTV News', title: 'Oasis Singer Liam Gallagher Arrested For Attacking Photographer', url: 'http://www.mtv.com/news/503357/oasis-singer-liam-gallagher-arrested-for-attacking-photographer/' },
-            title: 'Liam Gallagher Arrested for Attacking Photographer',
-            type: 'noteworthy',
-            year: '1998'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>The Masterplan</i> album sales as gold.</span>,
-            date: 'November 6',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Be Here Now Sales Certified as Gold by BPI',
-            type: 'certifications',
-            year: '1998'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>The Masterplan</i> album sales as platinum.</span>,
-            date: 'November 13',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Be Here Now Sales Certified as Platinum by BPI',
-            type: 'certifications',
-            year: '1998'
-        },
-        {
-            description: <span>After attending a celebrity party with bandmates Noel and Liam Gallagher to celebrate the opening of a Tommy Hilfiger shop in Mayfair, London, UK, Paul "Bonehead" Arthurs is arrested by police for drunk and disorderly behaviour.  Bonehead is later released without being charged.</span>,
-            date: 'February 23',
-            source: { name:'BBC News', title:'Oasis guitarist arrested', url:'http://news.bbc.co.uk/2/hi/entertainment/285558.stm'},
-            title: 'Paul "Bonehead" Arthurs Arrested in London',
-            type: 'noteworthy',
-            year: '1999'
-        },
-        {
-            description: <span>Oasis begin recording material for their fourth studio album in London, UK.  They also announce that they will be working with Mark "Spike" Stent as producer.</span>,
-            date: 'February 24',
-            source: { name: 'Official Oasis Site', title: 'Official News Updates', url: 'https://web.archive.org/web/19990428114830/http://www.oasisinet.com/news.htm' },
-            title: 'Oasis Begin Recording Standing on the Shoulder of Giants',
-            type: 'recordings',
-            year: '1999'
-        },
-        {
-            description: <span>Former Oasis drummer Tony McCarroll reaches an out of court settlement with the band before the trial was set to start.  In exchange for a £600,000 lump sum, he agrees to give up any future royalties.</span>,
-            date: 'March 3',
-            source: { name: 'The Independent', title: 'Sacked Oasis drummer in pounds 600,000 deal', url: 'https://www.independent.co.uk/news/sacked-oasis-drummer-in-pounds-600000-deal-1077972.html' },
-            title: 'Oasis and Tony McCarroll Reach Settlement',
-            type: 'noteworthy',
-            year: '1999'
-        },
-        {
-            description: <span>Oasis continue recording material for their fourth studio album at the Chateau De La Colle Noire, Montauroux, France with producer Mark "Spike" Stent.</span>,
-            date: 'April ?',
-            source: { name: 'The Hollywood Reporter', title: 'Inside Dior Parfums\' Intimate Gala at Christian Dior\'s French Chateau', url: 'https://www.hollywoodreporter.com/news/inside-dior-parfums-intimate-gala-892319' },
-            title: 'Recording Session for Standing on the Shoulder of Giants',
-            type: 'recordings',
-            year: '1999'
-        },
-        {
-            description: <span>Noel Gallagher performs as part of the Kosovo Aid Benefit Party at the HMV Forum in London, UK.  Noel performs a short acoustic set, including a cover of The Verve's <i>Bittersweet Symphony</i>.</span>,
-            date: 'May 30',
-            source: { name: 'Oasisinfo', title: 'May 30, 1999 - Kosovo Benefit Gig, Kentish Town Forum, London, England', url: 'http://oasisinfo.wikidot.com/1999-05-30' },
-            title: 'Noel Gallagher Performs at the Kosovo Aid Benefit Party',
-            type: 'gigs',
-            year: '1999'
-        },
-        {
-            description: <span>Original member and guitarist Paul "Bonehead" Arthurs announces that he is leaving the band he formed.</span>,
-            date: 'August 9',
-            source: { name: 'The Independent', title: 'Oasis guitarist \'Bonehead\' quits', url: 'https://www.independent.co.uk/news/oasis-guitarist-bonehead-quits-1111831.html' },
-            title: 'Paul "Bonehead" Arthurs Quits Oasis',
-            type: 'noteworthy',
-            year: '1999'
-        },
-        {
-            description: <span>Original member and bass guitarist Paul "Guigsy" McGuigan announces that he is leaving the band he formed.  He famously quits by sending a fax to the band and avoids phone calls from the Gallagher brothers in the following weeks.</span>,
-            date: 'August 25',
-            source: { name: 'The Independent', title: 'Now Guigsy deserts Oasis to \'spend time with his family\'', url: 'https://www.independent.co.uk/news/now-guigsy-deserts-oasis-to-spend-time-with-his-family-1115333.html' },
-            title: 'Paul "Guigsy" McGuigan Quits Oasis',
-            type: 'noteworthy',
-            year: '1999'
-        },
-        {
-            description: <span>Noel and Liam Gallagher appear at a hastily called press conference at Water Rats in London, UK to discuss the recent departures of original members Paul "Bonehead" Arthurs and Paul "Guigsy" McGuigan.</span>,
-            date: 'August 25',
-            source: { name: 'The Guardian', title: 'Oasis in crisis after second guitarist quits ', url: 'https://www.theguardian.com/uk/1999/aug/26/fiachragibbons1' },
-            title: 'Oasis Press Conference To Discuss Departed Members',
-            type: 'noteworthy',
-            year: '1999'
-        },
-        {
-            description: <span>Oasis announce the release date, track listing and title for their fourth studio album, <i>Standing On The Shoulder of Giants</i>.</span>,
-            date: 'October 29',
-            source: { name: 'Official Oasis Site', title: 'New LP Track Listing/December USA gigs announced', url: 'https://web.archive.org/web/19991127231307/http://www.oasisinet.com/prlp0200.htm' },
-            title: 'Oasis Announce New Album',
-            type: 'releases',
-            year: '1999'
-        },
-        {
-            description: <span>Colin "Gem" Archer, formerly of the band Heavy Stereo joins Oasis as the guitarist replacing departed member Paul "Bonehead" Arthurs.  His first appearance with the band is in the music video for <i>Go Let It Out</i>.</span>,
-            date: 'November 1',
-            source: { name: 'BBC Wear', title: 'Rise of a Northern Son', url: 'http://www.bbc.co.uk/wear/content/articles/2005/08/05/gem_archer_feature.shtml' },
-            title: 'Colin "Gem" Archer Joins Oasis',
-            type: 'noteworthy',
-            year: '1999'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Go Let It Out</i> in Surrey, UK.  It is the first appearance for "Gem" Archer with the band.  The video is directed by Nick Egan.</span>,
-            date: 'November 3',
-            source: { name: 'Live4Ever Forum', title: 'Oasis\' Videos', url: 'http://live4ever.proboards.com/post/393777/thread' },
-            title: 'Go Let It Out Music Video Filmed',
-            type: 'videos',
-            year: '1999'
-        },
-        {
-            description: <span>Andy Bell, formerly of the bands Ride and Hurricane #1 joins Oasis as the bass guitarist replacing departed member Paul "Guigsy" McGuigan.</span>,
-            date: 'November 12',
-            source: { name: 'BBC World News', title: 'Oasis recruit Ride man', url: 'http://news.bbc.co.uk/2/hi/entertainment/517724.stm' },
-            title: 'Andy Bell Joins Oasis',
-            type: 'noteworthy',
-            year: '1999'
-        },
-        {
-            description: <span>Alan McGee, the chairman of Creation Records and the man who signed Oasis, quits the label he founded.  This leads Oasis to launch their own record label, Big Brother Recordings Ltd.</span>,
-            date: 'November 25',
-            source: { name: 'The Guardian', title: '\'Oasis\' McGee quits label he created', url: 'https://www.theguardian.com/uk/1999/nov/26/paulkelso' },
-            title: 'Alan McGee Quits Creation Records',
-            type: 'noteworthy',
-            year: '1999'
-        },
-        {
-            description: <span>Oasis play their first gig at the First Union Center in Philadelphia, Pennsylvania, USA as part of the world tour in support of their upcoming fourth studio album <i>Standing on the Shoulder of Giants</i>.</span>,
-            date: 'December 3',
-            source: { name: 'Wikipedia', title: 'Standing on the Shoulder of Giants Tour', url: 'https://en.wikipedia.org/wiki/Standing_on_the_Shoulder_of_Giants_Tour' },
-            title: 'Standing On The Shoulder Of Giants World Tour Begins',
-            type: 'gigs',
-            year: '1999'
-        },
-        {
-            description: <span>After the dissolution of Creation Records, Oasis announces that they have formed their own record label, Big Brother, and will be releasing their upcoming album <i>Standing on the Shoulder of Giants</i> via the new label.</span>,
-            date: 'January 6',
-            source: { name: 'The Guardian', title: 'Oasis form their own record label', url: 'https://www.theguardian.com/uk/2000/jan/06/paulkelso' },
-            title: 'Oasis Form Big Brother Recordings',
-            type: 'noteworthy',
-            year: '2000'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Who Feels Love?</i> in Death Valley, California, USA.  The video is directed by Nick Egan.</span>,
-            date: 'January 10',
-            source: { name: 'Live4Ever Forum', title: 'Oasis\' Videos', url: 'http://live4ever.proboards.com/post/393777/thread' },
-            title: 'Who Feels Love? Music Video Filmed',
-            type: 'videos',
-            year: '2000'
-        },
-        {
-            description: <span>Noel Gallagher and Gem Archer appear on Steve Lamacq's Live Lounge program on BBC Radio 1 for an interview, performance and to select a cover based on fan submissions to be recorded later that week for the Evening Session program.  Noel ends up drawing <i>My Generation</i> by The Who.</span>,
-            date: 'January 17',
-            source: { name: 'Oasis Interviews', title: 'Noel Gallagher & Gem Archer - BBC Radio 1 - 17th January 2000', url: 'http://oasisinterviews.blogspot.com/2000/01/noel-gallagher-gem-archer-bbc-radio-1.html' },
-            title: 'Oasis Chooses Cover Song To Record From Fan Submissions on BBC Radio 1',
-            type: 'noteworthy',
-            year: '2000'
-        },
-        {
-            description: <span>Oasis record five songs for the Evening Session program on BBC Radio 1.  The session is notable as it includes the first performance of <i>My Generation</i> by the Who.</span>,
-            date: 'January 20',
-            source: { name: 'The Setlist Wiki', title: 'Oasis Setlist at Maida Value Studios, London, England', url: 'https://www.setlist.fm/setlist/oasis/2000/maida-vale-studios-london-england-bd7913e.html' },
-            title: 'Oasis Records For Evening Session on BBC Radio 1',
-            type: 'recordings',
-            year: '2000'
-        },
-        {
-            description: <span><i>Go Let It Out</i>, the first single from the album <i>Standing on the Shoulder of Giants</i>, is released.  It is the first release on Oasis' label Big Brother Recordings Ltd.  The single includes the b-sides <i>Let's All Make Believe</i> and <i>(As Long As They've Got) Cigarettes in Hell</i>.</span>,
-            date: 'February 7',
-            source: { name: 'Wikipedia', title: 'Got Let It Out', url: 'https://en.wikipedia.org/wiki/Go_Let_It_Out' },
-            title: 'Go Let It Out Released',
-            type: 'releases',
-            year: '2000'
-        },
-        {
-            description: <span>The music video for <i>Go Let It Out</i> is released.  The video is directed by Nick Egan.</span>,
-            date: 'February 7',
-            source: { name: 'IMDB', title: 'Oasis: Go Let It Out', url: 'https://www.imdb.com/title/tt5975554/' },
-            title: 'Go Let It Out Music Video Released',
-            type: 'videos',
-            year: '2000'
-        },
-        {
-            description: <span>Robbie Wiliams sends Noel Gallagher a funeral wreath reading "To Noel Gallagher, RIP.  Heard your latest album -- with deepest sympathy.  Robbie Williams."  It is the latest salvo in the ongoing feud between Williams and the Gallagher brothers.</span>,
-            date: 'February 25',
-            source: { name: 'The Free Library', title: 'Robbie sends a wreath to Noel over blunder call', url: 'https://www.thefreelibrary.com/Robbie+sends+a+wreath+to+Noel+over+blunder+call.-a060518274' },
-            title: 'Robbie Williams Sends Noel Gallagher Funeral Wreath',
-            type: 'noteworthy',
-            year: '2000'
-        },
-        {
-            description: <span><i>Standing on the Shoulder of Giants</i>, Oasis' fourth studio album, is released.  It is the first Oasis album released on their own label, Big Brother.  The album is the last to include original members Paul "Bonehead" Arthurs and Paul "Gigsy" McGuigan.</span>,
-            date: 'February 28',
-            source: { name: 'Wikipedia', title: 'Standing on the Shoulder of Giants', url: 'https://en.wikipedia.org/wiki/Standing_on_the_Shoulder_of_Giants' },
-            title: 'Standing on the Shoulder of Giants Released',
-            type: 'releases',
-            year: '2000'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Standing On The Shoulder of Giants</i> album sales as 2x platinum.</span>,
-            date: 'March 3',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Standing On The Shoulder of Giants Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '2000'
-        },
-        {
-            description: <span>During an acceptance speech for Best Video at the BRIT Awards in Earls Court, London, UK, Robbie Williams challenges Liam Gallagher to a televised boxing match.</span>,
-            date: 'March 3',
-            source: { name: 'MTV', title: 'Promoters Support Robbie VS Liam Match', url: 'http://www.mtv.com/news/1432793/promoters-support-robbie-vs-liam-match/' },
-            title: 'Robbie Williams Challenges Liam Gallagher to a Boxing Match',
-            type: 'noteworthy',
-            year: '2000'
-        },
-        {
-            description: <span><i>Who Feels Love?</i>, the second single from the album <i>Standing on the Shoulder of Giants</i>, is released on Big Brother.  The single includes the b-sides <i>One Way Road</i> and a cover of <i>Helter Skelter</i> originally by The Beatles.</span>,
-            date: 'April 17',
-            source: { name: 'Wikipedia', title: 'Who Feels Love?', url: 'https://en.wikipedia.org/wiki/Who_Feels_Love%3F' },
-            title: 'Who Feels Love? Released',
-            type: 'releases',
-            year: '2000'
-        },
-        {
-            description: <span>The music video for <i>Who Feels Love?</i> is released.  The video is directed by Nick Egan.</span>,
-            date: 'April 17',
-            source: { name: 'IMDB', title: 'Oasis: Who Feels Love?', url: 'https://www.imdb.com/title/tt5978810/' },
-            title: 'Who Feels Love? Music Video Released',
-            type: 'videos',
-            year: '2000'
-        },
-        {
-            description: <span>Noel Gallagher quits overseas touring with Oasis.  He announces that he will return for the final eight sell-out dates in the UK in July and August.</span>,
-            date: 'May 23',
-            source: { name: 'The Guardian', title: 'Oasis star quits world tour', url: 'https://www.theguardian.com/uk/2000/may/24/sarahhall' },
-            title: 'Noel Gallagher Quits World Tour',
-            type: 'noteworthy',
-            year: '2000'
-        },
-        {
-            description: <span>Oasis performs a gig at the Forum di Assago, Milan, Italy without Noel Gallagher.  The band drafts Matt Deighton of the band Mother Earth to fill in for the rest of the European leg of the tour.</span>,
-            date: 'May 30',
-            source: { name: 'The Irish Time', title: 'Oasis concert in Dublin to be among the last for Noel Gallagher', url: 'https://www.irishtimes.com/news/oasis-concert-in-dublin-to-be-among-the-last-for-noel-gallagher-1.274799' },
-            title: 'Oasis Plays Gig Without Noel Gallagher',
-            type: 'gigs',
-            year: '2000'
-        },
-        {
-            description: <span><i>Sunday Morning Call</i>, the third and final single from the album <i>Standing on the Shoulder of Giants</i>, is released on Big Brother.  The single includes the b-sides <i>Carry Us All</i> and <i>Full On</i>.</span>,
-            date: 'July 3',
-            source: { name: 'Wikipedia', title: 'Sunday Morning Call', url: 'https://en.wikipedia.org/wiki/Sunday_Morning_Call' },
-            title: 'Sunday Morning Call Released',
-            type: 'releases',
-            year: '2000'
-        },
-        {
-            description: <span>The music video for <i>Sunday Morning Call</i> is released.  The video is directed by Nick Egan.</span>,
-            date: 'July 3',
-            source: { name: 'IMDB', title: 'Oasis: Sunday Morning Call', url: 'https://www.imdb.com/title/tt5978892/' },
-            title: 'Sunday Morning Call Music Video Released',
-            type: 'videos',
-            year: '2000'
-        },
-        {
-            description: <span>Oasis plays two night at the old Wembley Stadium in London, UK.  The first night (of two) is released as <i>Familiar To Millions</i> later in the year.  Noel Gallagher would later remark that the second night was "the low point in Oasis."</span>,
-            date: 'July 21',
-            source: { name: 'BBC World News', title: 'Oasis\' Wembley woes', url: 'http://news.bbc.co.uk/2/hi/entertainment/844860.stm' },
-            title: 'Oasis Plays Wembley Stadium, London, UK - First Night',
-            type: 'gigs',
-            year: '2000'
-        },
-        {
-            description: <span>After bottles are thrown on stage during a performance, Oasis walk off stage at the Paleo Festival near L'Asse, Nyon, Switzerland after only six songs.</span>,
-            date: 'July 26',
-            source: { name: 'BBC World News', title: 'Oasis roll with the punches', url: 'http://news.bbc.co.uk/2/hi/uk_news/scotland/857105.stm' },
-            title: 'Oasis Storm Off Stage at Paleo Festival in Switzerland',
-            type: 'gigs',
-            year: '2000'
-        },
-        {	description: <span>Oasis ends their world tour in support of <i>Standing On The Shoulder Of Giants</i> at the Leeds Festival in Leeds, UK.  Noel Gallagher famously starts the gig by proclaiming that "this isn't the funeral" for Oasis despite the tumultuous tour.</span>,
-            date: 'August 28',
-            source: { name: 'Leeds Festival', title: 'Leeds 2000', url: 'https://www.leedsfestival.com/history/leeds-2000' },
-            title: 'Oasis Play the Final Gig of the Standing on the Shoulder of Giants World Tour',
-            type: 'gigs',
-            year: '2000'
-        },
-        {
-            description: <span>After three years of marriage, Liam Gallagher and Patsy Kensit divorce in London, UK.</span>,
-            date: 'September 24',
-            source: { name: 'BBC World News', title: 'All over for Liam and Patsy', url: 'http://news.bbc.co.uk/2/hi/entertainment/939810.stm' },
-            title: 'Liam Gallagher and Patsy Kensit Divorce',
-            type: 'noteworthy',
-            year: '2000'
-        },
-        {
-            description: <span>Along with other artists such as The Streophonics and Cornershop, Oasis performs covers of songs by John Lennon on the Channel 4 program <i>Shine On</i>.</span>,
-            date: 'September 30',
-            source: { name: 'IMDB', title: 'Shine On', url: 'https://www.imdb.com/title/tt0262765/' },
-            title: 'Oasis Perform as Part of Channel 4\'s Tribute to John Lennon',
-            type: 'noteworthy',
-            year: '2000'
-        },
-        {
-            description: <span>During his acceptance speech of Best Live At at the Q Awards at the Park Lane Hotel in London, UK, Liam Gallagher dedicates the award to Robbie Williams as part of their ongoing feud.  Instead of esclating, Williams would leave the ceremony early without confronting Gallagher.</span>,
-            date: 'October 31',
-            source: { name: 'BBC World News', title: 'Williams walks after Q quarrel', url: 'http://news.bbc.co.uk/2/hi/entertainment/1000246.stm' },
-            title: 'Liam Gallagher Thanks Robbie Williams at Q Awards',
-            type: 'noteworthy',
-            year: '2000'
-        },
-        {
-            description: <span><i>Familiar To Millions</i>, Oasis' first live album, is released on their own label Big Brother.  It is a recording of their concert at Wembley Stadium, London, UK on July 21, 2000.  It is released in audio and video formats.</span>,
-            date: 'November 13',
-            source: { name: 'Wikipedia', title: 'Familiar To Millions', url: 'https://en.wikipedia.org/wiki/Familiar_to_Millions' },
-            title: 'Familiar to Millions Released',
-            type: 'releases',
-            year: '2000'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Familiar To Millions</i> album sales as gold.</span>,
-            date: 'November 17',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Familiar To Millions Sales Certified as Gold by BPI',
-            type: 'certifications',
-            year: '2000'
-        },
-        {
-            description: <span>Noel Gallagher and Meg Matthews divorce after three years of marriage.  The divorce is granted based on the grounds of Noel's adultery with publicist Sara MacDonald.  Noel will later claim that he lied about when the relationship with MacDonald began in order to speed up the divorce proceedings.</span>,
-            date: 'January 19',
-            source: { name: 'BBC World News', title: 'Noel and Meg finally divorce', url: 'http://news.bbc.co.uk/2/hi/entertainment/1125485.stm' },
-            title: 'Noel Gallagher and Meg Matthews Divorce',
-            type: 'noteworthy',
-            year: '2001'
-        },
-        {
-            description: <span>Oasis begin a co-headlining tour of cities in North America with the Black Crowes at the Hard Rock Hotel in Paradise, Nevada, USA.</span>,
-            date: 'May 11',
-            source: { name: 'Wikipedia', title: 'The Tour of Brotherly Love', url: 'https://en.wikipedia.org/wiki/The_Tour_of_Brotherly_Love' },
-            title: 'The Tour of Brotherly Love Begins',
-            type: 'gigs',
-            year: '2001'
-        },
-        {
-            description: <span>The Recording Industry Association of America certifies <i>Definitely Maybe</i> album sales as platinum.</span>,
-            date: 'August 31',
-            source: { name: 'Recording Industry Association of America', title: 'Gold & Platinum Awards', url: 'https://www.riaa.com/gold-platinum/?tab_active=default-award&ar=Oasis&ti=&lab=&genre=&format=&date_option=release&from=1993-01-01&to=1998-12-31&award=&type=&category=&adv=SEARCH#search_section' },
-            title: 'Definitely Maybe Sales Certified as Platinum by RIAA',
-            type: 'certifications',
-            year: '2001'
-        },
-        {
-            description: <span>Oasis kicks off their 10th anniversary tour at Sheperd's Bush Empire in London, UK.  The tour is named the "10 Years of Noise and Confusion Tour".  It is a short six date tour of the UK.</span>,
-            date: 'October 7',
-            source: { name: 'BBC World News', title: 'Oasis: 10 years and rolling', url: 'http://news.bbc.co.uk/2/hi/entertainment/1580775.stm' },
-            title: '10 Years of Noise and Confusion Tour Begins',
-            type: 'gigs',
-            year: '2001'
-        },
-        {
-            description: <span>Oasis begins their world tour in support of their upcoming fifth studio album <i>Heathen Chemistry</i> at the Watford Coliseum in Watford, UK.</span>,
-            date: 'February 5',
-            source: { name: 'Wikipedia', title: 'Heathen Chemistry Tour', url: 'https://en.wikipedia.org/wiki/Heathen_Chemistry_Tour' },
-            title: 'Heathen Chemistry World Tour Begins',
-            type: 'gigs',
-            year: '2002'
-        },
-        {
-            description: <span>Oasis's forthcoming single <i>The Hindu Times</i> leaks on the internet a month before its official release.  The recording is a bootleg of the world premier of the song from a Norwegian radio station.</span>,
-            date: 'March 1',
-            source: { name: 'BBC News', title: 'New Oasis song leaked', url: 'http://news.bbc.co.uk/2/hi/entertainment/1849631.stm' },
-            title: 'The Hindu Times Leaks on Internet',
-            type: 'noteworthy',
-            year: '2002'
-        },
-        {
-            description: <span>After an incident occurs outside the Met Bar in London, UK on the evening of March 22nd, police question Liam to determine if an assault took place.</span>,
-            date: 'March 28',
-            source: { name: 'BBC News', title: 'Liam in police inquiry over \'assault\'', url: 'http://news.bbc.co.uk/2/hi/entertainment/1899318.stm' },
-            title: 'Liam Gallagher Questioned By Police After Incident in London Pub',
-            type: 'noteworthy',
-            year: '2002'
-        },
-        {
-            description: <span>The music video for <i>The Hindu Times</i> is released.  The video is directed by W.I.Z (Andrew Whiston).</span>,
-            date: 'April 4',
-            source: { name: 'IMDB', title: 'Oasis: The Hindu Times', url: 'https://www.imdb.com/title/tt7318296/' },
-            title: 'The Hindu Times Music Video Released',
-            type: 'videos',
-            year: '2002'
-        },
-        {
-            description: <span>Songs from Oasis's fifth studio album <i>Heathen Chemistry</i> leak on the Internet.  The British Phonographic Industry launches a probe into the leak, which it believes came from someone close to the band, not a journalist.</span>,
-            date: 'April 5',
-            source: { name: 'NME', title: 'New Oasis Tracks Leaked', url: 'https://www.nme.com/news/music/oasis-805-1384001' },
-            title: 'Heathen Chemistry Leaks on the Internet',
-            type: 'noteworthy',
-            year: '2002'
-        },
-        {
-            description: <span><i>The Hindu Times</i>, the first single from the album <i>Heathen Chemistry</i>, is released on Big Brother Recordings.  It includes the b-sides <i>Just Getting Older</i> and <i>Idler's Dream</i>.</span>,
-            date: 'April 15',
-            source: { name: 'Wikipedia', title: 'The Hindu Times', url: 'https://en.wikipedia.org/wiki/The_Hindu_Times' },
-            title: 'The Hindu Times Released',
-            type: 'releases',
-            year: '2002'
-        },
-        {
-            description: <span>Oasis headline the second night of the Coachella Valley Music and Arts Festival in Indio, California, USA.  The festival is attended by more than 55,000 people over two days.</span>,
-            date: 'April 28',
-            source: { name: 'MTV', title: 'Bjork, Oasis Headlining Coachella Festival', url: 'http://www.mtv.com/news/1452239/bjork-oasis-headlining-coachella-festival/' },
-            title: 'Oasis Play Coachella Valley Music and Arts Festival',
-            type: 'gigs',
-            year: '2002'
-        },
-        {
-            description: <span>The music video for <i>Stop Crying Your Heart Out</i> is released.  The video is directed by W.I.Z (Andrew Whiston).</span>,
-            date: 'May',
-            source: { name: 'IMDB', title: 'Oasis: Stop Crying Your Heart Out', url: 'https://www.imdb.com/title/tt7318286/' },
-            title: 'Stop Crying Your Heart Out Music Video Released',
-            type: 'videos',
-            year: '2002'
-        },
-        {
-            description: <span><i>Stop Crying Your Heart Out</i>, the second single from the album <i>Heathen Chemistry</i>, is released on Big Brother Recordings.  It includes the b-sides <i>Thank You For the Good Times</i> and <i>Shout It Out Loud</i>.</span>,
-            date: 'June 17',
-            source: { name: 'Wikipedia', title: 'Stop Crying Your Heart Out', url: 'https://en.wikipedia.org/wiki/Stop_Crying_Your_Heart_Out' },
-            title: 'Stop Crying Your Heart Out Released',
-            type: 'releases',
-            year: '2002'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Heathen Chemistry</i> album sales as platinum.</span>,
-            date: 'June 28',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Heathen Chemistry Sales Certified as Platinum by BPI',
-            type: 'certifications',
-            year: '2002'
-        },
-        {
-            description: <span><i>Heathen Chemistry</i>, Oasis' fifth studio album, is released on Big Brother Recordings.  The album is the first to include new members Colin "Gem" Archer and Andy Bell.</span>,
-            date: 'July 1',
-            source: { name: 'Wikipedia', title: 'Heathen Chemistry', url: 'https://en.wikipedia.org/wiki/Heathen_Chemistry' },
-            title: 'Heathen Chemistry Released',
-            type: 'releases',
-            year: '2002'
-        },
-        {
-            description: <span>Oasis play the first of three nights at Finsbury Park, London, UK.  The band play to 150,000 over the three day span and receive support from the Charlatans UK, The Coral, Proud Mary, Soundtrack of Our Lives and Black Rebel Motorcycle Club.</span>,
-            date: 'July 5',
-            source: { name: 'The Telegraph', title: 'Desert Oasis? Not while Liam\'s still breathing', url: 'https://www.telegraph.co.uk/culture/music/rockandjazzmusic/3579848/Desert-Oasis-Not-while-Liams-still-breathing.html' },
-            title: 'Oasis Play to 150,000 at Finsbury Park Over 3 Nights',
-            type: 'gigs',
-            year: '2002'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Little by Little</i> on Liverpool Street in London, UK.  The video stars actor Robert Carlyle and is directed by Max Giwa and Dania Pasquini (credited as Max & Dania).</span>,
-            date: 'July 8',
-            source: { name: 'Live4Ever Forum', title: 'Oasis\' Videos', url: 'http://live4ever.proboards.com/post/393777/thread' },
-            title: 'Little By Little Music Video Filmed',
-            type: 'videos',
-            year: '2002'
-        },
-        {
-            description: <span>The music video for <i>Little By Little</i> is released.  The video is directed by "Max & Dania" (Max Giwa and Dania Pasquini).</span>,
-            date: 'August',
-            source: { name: 'IMDB', title: 'Oasis: Little By Little', url: 'https://www.imdb.com/title/tt7318276/' },
-            title: 'Little By Little Music Video Released',
-            type: 'videos',
-            year: '2002'
-        },
-        {
-            description: <span>Noel Gallagher is almost left behind at Heathrow Airport in London, UK after he is unable to find his passport.  The US embassy in the UK helps to rush through a working visa so that Noel may join the tour in the US.  Years later, Liam Gallagher claims that Noel's girlfriend Sara MacDonald stole Noel's passport to prevent him from leaving home.  This story has never been corroborated.</span>,
-            date: 'August 1',
-            source: { name: 'BBC World News', title: 'Lost passport threatens Oasis tour', url: 'http://news.bbc.co.uk/2/hi/entertainment/2168404.stm' },
-            title: 'Oasis Tour In Doubt After Noel Gallagher Loses Passport',
-            type: 'noteworthy',
-            year: '2002'
-        },
-        {
-            description: <span>After just four songs, Liam Gallagher walks off stage during Oasis' first gig of the <i>Heathen Chemistry</i> tour at the Pompano Beach Amphitheater in Fort Lauderdale, Florida, USA.  The band states that Liam is suffering from a throat infection that also forced the band to cancel a gig 10 days earlier in Spain.</span>,
-            date: 'August 2',
-            source: { name: 'BBC World News', title: 'Oasis Liam walks off stage', url: 'http://news.bbc.co.uk/2/hi/entertainment/2173577.stm' },
-            title: 'Liam Gallagher Walks Off Stage',
-            type: 'gigs',
-            year: '2002'
-        },
-        {
-            description: <span>Noel Gallagher, Andy Bell and Jay Darlington are injured during the US tour after a head-on car crash involving their taxi and another car in Indianapolis, Indiana, USA.  A couple tour dates are cancelled before Oasis continues their US tour in support of their fifth studio album <i>Heathen Chemistry</i>.</span>,
-            date: 'August 7',
-            source: { name: 'The Telegraph', title: 'Noel Gallagher injured in car crash', url: 'https://www.telegraph.co.uk/news/worldnews/northamerica/usa/1403913/Noel-Gallagher-injured-in-car-crash.html' },
-            title: 'Oasis Members Injured in Car Accident',
-            type: 'noteworthy',
-            year: '2002'
-        },
-        {
-            description: <span>The music video for <i>She Is Love</i> is released.  The video is directed by Rachel Thomas.</span>,
-            date: 'September',
-            source: { name: 'IMDB', title: 'Oasis: She Is Love', url: 'https://www.imdb.com/title/tt6023876/' },
-            title: 'She Is Love Music Video Released',
-            type: 'videos',
-            year: '2002'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Heathen Chemistry</i> album sales as 2x platinum.</span>,
-            date: 'September 13',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Heathen Chemistry Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '2002'
-        },
-        {
-            description: <span><i>Little by Little</i>, the third single from the album <i>Heathen Chemistry</i>, is released on Big Brother Recordings featuring the double A-sides <i>Little By Little</i> and <i>She is Love</i>.  It includes the b-side cover of <i>My Generation</i> originally performed by the Who.</span>,
-            date: 'September 23',
-            source: { name: 'Wikipedia', title: 'Little By Little (Oasis song)', url: 'https://en.wikipedia.org/wiki/Little_by_Little_(Oasis_song)' },
-            title: 'Little By Little/She Is Love Released',
-            type: 'releases',
-            year: '2002'
-        },
-        {
-            description: <span>Noel Gallagher and Chris Martin perform together at the London Astoria in London, UK as part of the "Fair Play" charity event to raise money and awareness for Oxfam.  The concert is organized by Emily Eavis, daugher of the Glastonbury Festival founder Michael Eavis.  Noel and Chris perform <i>Stop Me If You Think You've Heard this One Before</i> by The Smiths, <i>Yellow</i> by Coldplay, <i>Live Forever</i> and <i>Whatever</i> by Oasis together.</span>,
-            date: 'October 29',
-            source: { name: 'NME', title: 'Noel Gallagher and Chris Martin team up for duets in London', url: 'http://nme.com/news/103359.htm' },
-            title: 'Noel Gallagher and Chris Martin Duet for Charity in London',
-            type: 'noteworthy',
-            year: '2002'
-        },
-        {
-            description: <span>Liam Gallagher, Alan White, Phil Smith and members of the bands security staff are arrested after a brawl with patrons at the Bayerischer Hof hotel in Munich, Germany.  Gallagher would lose two teeth in the ensuing brawl with police.</span>,
-            date: 'December 2',
-            source: { name: 'The Independent', title: 'Oasis singer loses teeth and faces jail after \'drunken fight with German police\'', url: 'https://www.independent.co.uk/arts-entertainment/music/news/oasis-singer-loses-teeth-and-faces-jail-after-drunken-fight-with-german-police-134411.html' },
-            title: 'Oasis Involved in Brawl at German Hotel',
-            type: 'noteworthy',
-            year: '2002'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Songbird</i> in London, UK.  The video is directed by Dick Carruthers.</span>,
-            date: 'December 7',
-            source: { name: 'Live4Ever Forum', title: 'Oasis\' Videos', url: 'http://live4ever.proboards.com/post/393777/thread' },
-            title: 'Songbird Music Video Filmed',
-            type: 'videos',
-            year: '2002'
-        },
-        {
-            description: <span><i>Songbird</i>, the fourth single from the album <i>Heathen Chemistry</i>, is released on Big Brother Recordings.  It includes the b-sides <i>(You've Got) The Heart of a Star</i> and a live version of <i>Columbia</i>.</span>,
-            date: 'February 3',
-            source: { name: 'Wikipedia', title: 'Songbird (Oasis song)', url: 'https://en.wikipedia.org/wiki/Songbird_(Oasis_song)' },
-            title: 'Songbird Released',
-            type: 'releases',
-            year: '2003'
-        },
-        {
-            description: <span>The music video for <i>Songbird</i> is released.  The video is directed by Dick Carruthers.</span>,
-            date: 'February 3',
-            source: { name: 'IMDB', title: 'Oasis: Songbird', url: 'https://www.imdb.com/title/tt7078214/' },
-            title: 'Songbird Music Video Released',
-            type: 'videos',
-            year: '2003'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Heathen Chemistry</i> album sales as 3x platinum.</span>,
-            date: 'February 14',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Heathen Chemistry Sales Certified as 3x Platinum by BPI',
-            type: 'certifications',
-            year: '2003'
-        },
-        {
-            description: <span>Unknown at the time, Alan White performs with Oasis for the last time during a performance at the Arena Berlin in Berlin, Germany.  Alan would be sacked by the band before their next world tour.</span>,
-            date: 'March 12',
-            source: { name: 'Wikipedia', title: 'Heathen Chemistry Tour', url: 'https://en.wikipedia.org/wiki/Heathen_Chemistry_Tour' },
-            title: 'Alan White\'s Last Gig with Oasis',
-            type: 'gigs',
-            year: '2003'
-        },
-        {
-            description: <span>Former Oasis drummer Tony McCarroll loses his appeal over the handling the the settlement from his previous legal action against the band.</span>,
-            date: 'April 1',
-            source: { name: 'BBC World News', title: 'Former Oasis drummer loses appeal', url: 'http://news.bbc.co.uk/2/hi/entertainment/2906365.stm' },
-            title: 'Ex-drummer Tony McCarroll Loses Appeal',
-            type: 'noteworthy',
-            year: '2003'
-        },
-        {
-            description: <span>Playing an unannounced gig with Gem Archer and Terry Kirkbride, Noel Gallagher performs a short set at the Zanzibar Club in Liverpool, UK.  The show is notable because Noel plays <i>Stop the Clocks</i> for the first and only time live.  Other songs played include <i>Talk Tonight</i>, <i>Wonderwall</i>, <i>Don't Look Back In Anger</i> and <i>Married With Children</i>.</span>,
-            date: 'May 3',
-            source: { name: 'Getintothis', title: 'Lost Liverpool #19: The Bandwagon, The Zanzibar and famous nights with Noel Gallagher ', url: 'http://www.getintothis.co.uk/2017/05/lost-liverpool-19-bandwagon-zanzibar-famous-nights-noel-gallagher/' },
-            title: 'Noel Gallagher Performs "Stop the Clocks" Live',
-            type: 'gigs',
-            year: '2003'
-        },
-        {
-            description: <span>During recording sessions with dance act Death In Vegas, it is announced that Alan White has been sacked by the band.  No replacement is named.</span>,
-            date: 'January 16',
-            source: { name: 'BBC World News', title: 'Oasis drummer Alan White sacked', url: 'http://news.bbc.co.uk/2/hi/entertainment/3407719.stm' },
-            title: 'Alan White Sacked as Drummer',
-            type: 'noteworthy',
-            year: '2004'
-        },
-        {
-            description: <span>Oasis stop working with producers Death In Vegas for their sixth studio album.  They had spent time at Sawmill Studios in Cornwall, UK but decide instead of get a different producer and use their own studios to finish recording demos for the upcoming album.</span>,
-            date: 'March 9',
-            source: { name: 'NME', title: 'HE BANGS THE DRUMS!', url: 'https://www.nme.com/news/music/oasis-622-1365369' },
-            title: 'Oasis Abandon Death In Vegas Recording Sessions',
-            type: 'recordings',
-            year: '2004'
-        },
-        {
-            description: <span>Zak Starkey agrees to join Oasis, replacing Alan White as the band's drummer.</span>,
-            date: 'May 11',
-            source: { name: 'Evening Standard', title: 'Beatle heir joins Oasis ', url: 'https://www.standard.co.uk/news/beatle-heir-joins-oasis-7289180.html' },
-            title: 'Zak Starkey Joins Oasis',
-            type: 'noteworthy',
-            year: '2004'
-        },
-        {
-            description: <span>Oasis play a Glastonbury Festival warm-up gig at Poole Lighthouse, Poole, UK.  It is the first live performance with Zak Starkey as drummer and features the live debut of two new songs: <i>A Bell Will Ring</i> and <i>The Meaning of Soul</i>.</span>,
-            date: 'June 23',
-            source: { name: 'NME', title: 'Oasis Return!', url: 'https://www.nme.com/news/music/oasis-634-1366134' },
-            title: 'Zak Starkey Debuts Live With Oasis',
-            type: 'gigs',
-            year: '2004'
-        },
-        {
-            description: <span>Oasis perform a headline set at the Glastonbury Festival at Worthy Farm, Shepton Mallet, Pilton, UK.  It is the second time the band has been a headlining act at the festival (the other was in 1995) and only their third appearance overall at the festival.</span>,
-            date: 'June 25',
-            source: { name: 'Oasis', title: 'Oasis fail to surprise Glastonbury', url: 'http://news.bbc.co.uk/2/hi/entertainment/3842947.stm' },
-            title: 'Oasis Headline Glastonbury Festival for the Second Time',
-            type: 'gigs',
-            year: '2004'
-        },
-        {
-            description: <span>To celebrate the 10th anniversary of its release, <i>Definitely Maybe</i> is re-released with an accompanying film looking back at the recording of the album.</span>,
-            date: 'September 6',
-            source: { name: 'Wikipedia', title: 'Definitely Maybe', url: 'https://en.wikipedia.org/wiki/Definitely_Maybe#DVD' },
-            title: 'Definitely Maybe Re-Released',
-            type: 'releases',
-            year: '2004'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 14x platinum.</span>,
-            date: 'September 24',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 14x Platinum by BPI',
-            type: 'certifications',
-            year: '2004'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Definitely Maybe</i> album sales as 7x platinum.</span>,
-            date: 'January 30',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Definitely Maybe Sales Certified as 7x Platinum by BPI',
-            type: 'certifications',
-            year: '2005'
-        },
-        {
-            description: <span>Oasis announce that their sixth studio album will be titled <i>Don't Believe The Truth</i> and will be released in May.  The first single will be <i>Lyla</i>.</span>,
-            date: 'March 7',
-            source: { name: 'NME', title: 'Oasis Reveal The Whole \'Truth\'', url: 'https://www.nme.com/news/music/oasis-618-1364419' },
-            title: 'Oasis Announce New Album and Single',
-            type: 'noteworthy',
-            year: '2005'
-        },
-        {
-            description: <span>Oasis film the music video for <i>Lyla</i> at the Round Chapel in Hackney, East London, UK.  The video features a live performance of the song and was directed by Tim Qualtrough.</span>,
-            date: 'March 18',
-            source: { name: 'Kathy\'s Zak Starkey Site', title: 'Oasis 2005 TV and Radio Appearances', url: 'https://www.kathyszaksite.com/oasis20053_tvradio.html' },
-            title: 'Lyla Music Video Filmed',
-            type: 'videos',
-            year: 2005
-        },
-        {
-            description: <span> Oasis begins its world tour in support of its forthcoming sixth studio album <i>Don't Believe The Truth</i> by playing to 2,000 fans at the London Astoria, London, UK.</span>,
-            date: 'May 10',
-            source: { name: 'Wikipedia', title: 'Don\'t Believe The Truth Tour', url: 'https://en.wikipedia.org/wiki/Don%27t_Believe_the_Truth_Tour' },
-            title: 'Don\'t Believe The Truth World Tour Begins',
-            type: 'gigs',
-            year: '2005'
-        },
-        {
-            description: <span>First single from the album <i>Don't Believe The Truth</i> released on Big Brother Recordings.  It includes the b-sides <i>Eyeball Tickler</i> and <i>Won't Let You Down</i>.</span>,
-            date: 'May 16',
-            source: { name: 'Wikipedia', title: 'Lyla (song)', url: 'https://en.wikipedia.org/wiki/Lyla_(song)' },
-            title: 'Lyla Released',
-            type: 'releases',
-            year: '2005'
-        },
-        {
-            description: <span>The music video for <i>Lyla</i> is released.  The video is directed by Tim Qualtrough.</span>,
-            date: 'May 16',
-            source: { name: 'IMDB', title: 'Oasis: Lyla', url: 'https://www.imdb.com/title/tt5971262/' },
-            title: 'Lyla Music Video Released',
-            type: 'videos',
-            year: '2005'
-        },
-        {
-            description: <span>Oasis film the music video for <i>The Importance of Being Idle</i> in Greenwich, London, UK.  The video features Rhys Ifans and is directed by Dawn Shadforth.</span>,
-            date: 'May 17',
-            source: { name: 'Kathy\'s Zak Starkey Site', title: 'Oasis 2005 TV and Radio Appearances', url: 'https://www.kathyszaksite.com/oasis20053_tvradio.html' },
-            title: 'The Importance of Being Idle Music Video Filmed',
-            type: 'videos',
-            year: 2005
-        },
-        {
-            description: <span><i>Don't Believe the Truth</i>, Oasis' sixth studio album, is released on Big Brother Recordings.  The album is the first to include new members Zak Starkey on drums.</span>,
-            date: 'May 30',
-            source: { name: 'Wikipedia', title: 'Don\'t Believe the Truth', url: 'https://en.wikipedia.org/wiki/Don%27t_Believe_the_Truth' },
-            title: 'Don\'t Believe The Truth Released',
-            type: 'releases',
-            year: '2005'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Don't Believe The Truth</i> album sales as 2x platinum.</span>,
-            date: 'June 17',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Don\'t Believe The Truth Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '2005'
-        },
-        {
-            description: <span>Oasis plays Madison Square Garden in New York, New York, USA for the first time as part of its tour for <i>Don't Believe The Truth</i>.</span>,
-            date: 'June 22',
-            source: { name: 'New York Times', title: 'Band of Brothers, but Without Much Brotherly Love', url: 'https://www.nytimes.com/2005/06/24/arts/music/band-of-brothers-but-without-much-brotherly-love.html' },
-            title: 'Oasis Play Madison Square Garden',
-            type: 'gigs',
-            year: '2005'
-        },
-        {
-            description: <span>Oasis plays three sell-out shows at the home of Manchester City Football Club (City of Manchester Stadium) in Manchester, UK.  The July 2nd performance is filmed and included as part of the <i>Lord Don't Slow Me Down</i> "rockumentary".</span>,
-            date: 'June 30',
-            source: { name: 'Wikipedia', title: 'Lord Don\'t Slow Me Down', url: 'https://en.wikipedia.org/wiki/Lord_Don%27t_Slow_Me_Down' },
-            title: 'City of Manchester Stadium, Manchester, UK (First Night)',
-            type: 'gigs',
-            year: '2005'
-        },
-        {
-            description: <span><i>The Importance of Being Idle</i>, the second single from the album <i>Don't Believe The Truth</i>, is released on Big Brother Recordings.  It includes the b-sides <i>Pass Me Down the Wine</i> and <i>The Quiet Ones</i>.</span>,
-            date: 'August 22',
-            source: { name: 'Wikipedia', title: 'The Importance of Being Idle (song)', url: 'https://en.wikipedia.org/wiki/The_Importance_of_Being_Idle_(song)' },
-            title: 'The Importance of Being Idle Released',
-            type: 'releases',
-            year: '2005'
-        },
-        {
-            description: <span>The music video for <i>The Importance of Being Idle</i> is released.  The video is directed by Dawn Shadforth.</span>,
-            date: 'August 22',
-            source: { name: 'IMDB', title: 'Oasis: The Importance of Being Idle', url: 'https://www.imdb.com/title/tt5971324/' },
-            title: 'The Importance of Being Idle Music Video Released',
-            type: 'videos',
-            year: '2005'
-        },
-        {
-            description: <span><i>Let There Be Love</i>, the third single from the album <i>Don't Believe The Truth</i>, is released on Big Brother Recordings.  It includes the b-sides <i>Sittin' Here In Silence (On My Own)</i> and a live version of <i>Rock 'n' Roll Star</i>.</span>,
-            date: 'November 28',
-            source: { name: 'Wikipedia', title: 'Let There Be Love (Oasis song)', url: 'https://en.wikipedia.org/wiki/Let_There_Be_Love_(Oasis_song)' },
-            title: 'Let There Be Love',
-            type: 'releases',
-            year: '2005'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Don't Believe The Truth</i> album sales as 3x platinum.</span>,
-            date: 'January 6',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Don\'t Believe The Truth Sales Certified as 3x Platinum by BPI',
-            type: 'certifications',
-            year: '2006'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Familiar To Millions</i> album sales as platinum.</span>,
-            date: 'March 4',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Familiar To Millions Sales Certified as Platinum by BPI',
-            type: 'certifications',
-            year: '2006'
-        },
-        {
-            description: <span>The music video for <i>The Masterplan</i> is released.  The video is directed by Greg Fay and Ben Jones.</span>,
-            date: 'October 21',
-            source: { name: 'IMDB', title: 'Oasis: The Masterplan', url: 'https://www.imdb.com/title/tt7079280/' },
-            title: 'The Masterplan Music Video Released',
-            type: 'videos',
-            year: '2006'
-        },
-        {
-            description: <span><i>Stop The Clocks</i>, a retrospective double album of songs chosen by Noel Gallagher from the band's catalog to date, is released on Big Brother Recordings.</span>,
-            date: 'November 20',
-            source: { name: 'Wikipedia', title: 'Stop the Clocks', url: 'https://en.wikipedia.org/wiki/Stop_the_Clocks' },
-            title: 'Stop the Clocks Released',
-            type: 'releases',
-            year: '2006'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Stop The Clocks</i> album sales as 2x platinum.</span>,
-            date: 'December 1',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Stop The Clocks Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '2006'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Stop The Clocks</i> album sales as 3x platinum.</span>,
-            date: 'December 15',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Stop The Clocks Sales Certified as 3x Platinum by BPI',
-            type: 'certifications',
-            year: '2006'
-        },
-        {
-            description: <span>As part of a semi-acoustic tour, Noel Gallagher and Gem Archer become the first members of Oasis to play a live gig in Russia.  The show takes place at the Club B1 in Moscow.</span>,
-            date: 'March 23',
-            source: { name: 'NME', title: 'Noel Gallagher plays first Russian gig', url: 'https://www.nme.com/news/music/noel-gallagher-372-1344260' },
-            title: 'Noel Gallagher and Gem Archer Play Gig in Moscow, Russia',
-            type: 'gigs',
-            year: '2007'
-        },
-        {
-            description: <span><i>Lord Don't Slow Me Down</i>, a download-only single, is released in promotion of the upcoming "rockumentary" of the same name.  The single included the tracks <i>The Meaning Of Soul</i> (Live at City of Manchester Stadium '05) and <i>Don't Look Back In Anger</i> (Live at City of Manchester Stadium '05).</span>,
-            date: 'October 21',
-            source: { name: 'Wikipedia', title: 'Lord Don\'t Slow Me Down (song)', url: 'https://en.wikipedia.org/wiki/Lord_Don%27t_Slow_Me_Down_(song)' },
-            title: 'Lord Don\'t Slow Me Down (Song) Released',
-            type: 'releases',
-            year: '2007'
-        },
-        {
-            description: <span>Oasis release a "rockumentary" about the world tour in support of the album <i>Don't Believe The Truth</i>, which took place between May 2005 and March 2006.  The film is directed by Baillie Walsh.</span>,
-            date: 'October 29',
-            source: { name: 'Wikipedia', title: 'Lord Don\'t Slow Me Down', url: 'https://en.wikipedia.org/wiki/Lord_Don%27t_Slow_Me_Down' },
-            title: 'Lord Don\'t Slow Me Down (Film) Released',
-            type: 'releases',
-            year: '2007'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Stop The Clocks</i> album sales as 3x platinum.</span>,
-            date: 'November 2',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Stop The Clocks Sales Certified as 3x Platinum by BPI',
-            type: 'certifications',
-            year: '2007'
-        },
-        {
-            description: <span>Oasis begin recording their seventh studio album at <a href="https://www.abbeyroad.com/" target="_blank" rel="noopener noreferrer">Abbey Road Studios</a> in London, UK.  It is the band's first time recording at Abbey Road Studio since the aborted sessions for <i>Be Here Now</i> in 1996.</span>,
-            date: 'November 5',
-            source: { name: 'Gigwise', title: 'Oasis Begin Recording New Album at Abbey Road', url: 'https://www.gigwise.com/news/38173/' },
-            title: 'Oasis Begin Recording Dig Out Your Soul at Abbey Road Studios',
-            type: 'recordings',
-            year: '2007'
-        },
-        {
-            description: <span>Liam Gallagher and long time girlfriend Nicole Appleton (of the all-girl band All Saints) marry in a secret ceremony at the Westminster Register Office in London, UK.</span>,
-            date: 'February 14',
-            source: { name: 'NME', title: 'Oasis\' Liam Gallagher marries Nicole Appleton ', url: 'https://www.nme.com/news/music/oasis-497-1327219' },
-            title: 'Liam Gallagher and Nicole Appleton Wed',
-            type: 'noteworthy',
-            year: '2008'
-        },
-        {
-            description: <span>Zak Starkey quits Oasis, citing issues with committing to the upcoming world tour to support their new album <i>Dig Out Your Soul</i>.</span>,
-            date: 'May 12',
-            source: { name: 'NME', title: 'Oasis member quits ', url: 'https://www.nme.com/news/music/oasis-488-1327014' },
-            title: 'Zak Starkey Quits Oasis',
-            type: 'noteworthy',
-            year: '2008'
-        },
-        {
-            description: <span>Chris Sharrock replaces Zak Starkey as the drummer in Oasis.</span>,
-            date: 'May 15',
-            source: { name: 'Gigwise', title: 'Robbie Williams Drummer to Join Oasis', url: 'https://www.gigwise.com/news/43163/' },
-            title: 'Chris Sharrock Joins Oasis',
-            type: 'noteworthy',
-            year: '2008'
-        },
-        {
-            description: <span>Oasis renews its contract with Sony BMG for worldwide marketing and distribution for three more album.</span>,
-            date: 'June 20',
-            source: { name: 'Reuters', title: 'Oasis extend deal at Sony BMG', url: 'https://www.reuters.com/article/music-oasis-dc/oasis-extend-deal-at-sony-bmg-idUSN2038405220080620' },
-            title: 'Oasis Extend Deal with Sony BMG',
-            type: 'noteworthy',
-            year: '2008'
-        },
-        {
-            description: <span>Oasis announce that they are set to release their seventh studio album titled <i>Dig Out Your Soul</i> in October.  The first single will be <i>The Shock Of The Lightning</i>.</span>,
-            date: 'June 24',
-            source: { name: 'Reuters', title: 'Oasis announce new album release date', url: 'https://uk.reuters.com/article/uk-oasis-idUKL2580153820080625' },
-            title: 'Oasis Announce New Album and Single Details',
-            type: 'noteworthy',
-            year: '2008'
-        },
-        {
-            description: <span>The music video for <i>The Shock of the Lightning</i> is released.  The video is directed by Julian Gibbs and Julian House.</span>,
-            date: 'August 25',
-            source: { name: 'IMDB', title: 'Oasis: The Shock of the Lightning', url: 'https://www.imdb.com/title/tt5945798/' },
-            title: 'The Shock of the Lightning Music Video Released',
-            type: 'videos',
-            year: '2008'
-        },
-        {
-            description: <span>Oasis begin their world tour in support of their new album <i>Dig Out Your Soul</i> at the WaMu Theater in Seattle, Washington, USA.</span>,
-            date: 'August 26',
-            source: { name: 'Wikipedia', title: 'Dig Out Your Soul Tour', url: 'https://en.wikipedia.org/wiki/Dig_Out_Your_Soul_Tour' },
-            title: 'Dig Out Your Soul World Tour Begins',
-            type: 'gigs',
-            year: '2008'
-        },
-        {
-            description: <span>Noel Gallagher is attacked on stage by a fan while performing the song <i>Morning Glory</i> at the V Festival in Toronto, Ontario, Canada.  Noel suffers a broken rib in the attack.  The tour resumes a few nights later in New York, New York, USA.</span>,
-            date: 'September 7',
-            source: { name: 'Reuters', title: 'Oasis singer Noel Gallagher attacked on stage', url: 'https://www.reuters.com/article/us-oasis-idUSN0848540120080909' },
-            title: 'Noel Gallagher Attacked on Stage',
-            type: 'noteworthy',
-            year: '2008'
-        },
-        {
-            description: <span><i>The Shock of the Lightning</i>, the first single from the album <i>Dig Out Your Soul</i>, is released on Big Brother Recordings.  It includes the a Chemical Brothers remix of <i>Falling Down</i> as the b-side.</span>,
-            date: 'September 29',
-            source: { name: 'Wikipedia', title: 'The Shock of the Lightning', url: 'https://en.wikipedia.org/wiki/The_Shock_of_the_Lightning' },
-            title: 'The Shock of the Lightning Released',
-            type: 'releases',
-            year: '2008'
-        },
-        {
-            description: <span>Oasis releases their seventh and final studio album, <i>Dig Out Your Soul</i>.  It is the fourth released on their own label, Big Brother.</span>,
-            date: 'October 6',
-            source: { name: 'Wikipedia', title: 'Dig Out Your Soul', url: 'https://en.wikipedia.org/wiki/Dig_Out_Your_Soul' },
-            title: 'Dig Out Your Soul Released',
-            type: 'releases',
-            year: '2008'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Dig Out Your Soul</i> album sales as platinum.</span>,
-            date: 'October 10',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified ', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Dig Out Your Soul Sales Certified as Platinum by BPI',
-            type: 'certifications',
-            year: '2008'
-        },
-        {
-            description: <span>The music video for <i>I'm Outta Time</i> is released.  The video is directed by Intro.</span>,
-            date: 'November 6',
-            source: { name: 'IMDB', title: 'Oasis: I\'m Outta Time', url: 'https://www.imdb.com/title/tt5945822/' },
-            title: 'I\'m Outta Time Music Video Released',
-            type: 'videos',
-            year: '2008'
-        },
-        {
-            description: <span><i>I'm Outta Time</i>, the second single from the album <i>Dig Out Your Soul</i>, is released on Big Brother Recordings.  It includes a remix of <i>I'm Outta Time</i> and a Jagz Kooner remix of <i>The Shock Of The Lightning</i> as b-sides.</span>,
-            date: 'December 1',
-            source: { name: 'Wikipedia', title: 'I\'m Outta Time', url: 'https://en.wikipedia.org/wiki/I%27m_Outta_Time' },
-            title: 'I\'m Outta Time Released',
-            type: 'releases',
-            year: '2008'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Dig Out Your Soul</i> album sales as 2x platinum.</span>,
-            date: 'December 12',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Dig Out Your Soul Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '2008'
-        },
-        {
-            description: <span>The music video for <i>Falling Down</i> is released.  The video is directed by W.I.Z. (Andrew Whiston).</span>,
-            date: 'February 5',
-            source: { name: 'IMDB', title: 'Oasis: Falling Down', url: 'https://www.imdb.com/title/tt5945798/' },
-            title: 'Falling Down Music Video Released',
-            type: 'videos',
-            year: '2009'
-        },
-        {
-            description: <span>The Chinese government deems Oasis "unsuitable" to play in the country after discovering Noel Gallagher's performance during a "Free Tibet" concert in New York, USA in 1997.  The band was set to play in China for the first time in April.</span>,
-            date: 'February 28',
-            source: { name: 'The Guardian', title: 'Authorities ban Oasis from playing in China', url: 'https://www.theguardian.com/music/2009/mar/03/oasis-china-banned-concerts-tibet' },
-            title: 'Oasis Banned From Performing in China',
-            type: 'noteworthy',
-            year: '2009'
-        },
-        {
-            description: <span><i>Falling Down</i>, the third single from the album <i>Dig Out Your Soul</i>, is released on Big Brother Recordings.  It includes b-side <i>Those Swollen Hand Blues</i>.  It would become the band's final single.</span>,
-            date: 'March 9',
-            source: { name: 'Wikipedia', title: 'Falling Down (Oasis song)', url: 'https://en.wikipedia.org/wiki/Falling_Down_(Oasis_song)' },
-            title: 'Falling Down Released',
-            type: 'releases',
-            year: '2009'
-        },
-        {
-            description: <span>Oasis play their first gig on the African continent when they play the Riversands Farm in Johannesburg, South Africa.</span>,
-            date: 'April 10',
-            source: { name: 'Channel24', title: 'Coca-Cola Zero Fest', url: 'https://www.channel24.co.za/Music/FeaturesInterviews/Coca-Cola-Zero-Fest-20090122' },
-            title: 'Oasis Play First Gig in Africa',
-            type: 'gigs',
-            year: '2009'
-        },
-        {
-            description: <span>The first night of Oasis's 3 night performance at Heaton Park in Manchester, UK is interrupted when a generator fails.  After a lengthy delay, the band resumes the concert, not before Noel Gallagher tells the crowd that "this is a free gig".</span>,
-            date: 'June 4',
-            source: { name: 'The Guardian', title: 'Oasis offer fans refund for gig shambles', url: 'https://www.theguardian.com/music/2009/jun/08/oasis-offer-fans-refund' },
-            title: 'Oasis Offer Fans Refunds After Gig Interrupted Due to Technical Problems',
-            type: 'gigs',
-            year: '2009'
-        },
-        {
-            description: <span>Download-only EP released featuring the additional tracks <i>I Believe In All</i> and remix of <i>(Get Off Your) High Horse Lady</i>.</span>,
-            date: 'July 13',
-            source: { name: 'Wikipedia', title: 'Boy with the Blues', url: 'https://en.wikipedia.org/wiki/Boy_with_the_Blues' },
-            title: 'Boy with the Blues Download EP Released',
-            type: 'releases',
-            year: '2009'
-        },
-        {
-            description: <span>After promising refunds for technical problems during the band's June 7th Heaton Park concert, Oasis send out cheques for around £40 featuring both Noel and Liam Gallagher's signature.</span>,
-            date: 'July 13',
-            source: { name: 'The Guardian', title: 'Oasis refund fans almost £1m', url: 'https://www.theguardian.com/music/2009/jul/13/oasis-refund-fans' },
-            title: 'Oasis Issue Refunds for Heaton Park Concert',
-            type: 'noteworthy',
-            year: '2009'
-        },
-        {
-            description: <span>What will become their last gig as a band, Oasis play a V Festival slot at Weston Park, Stafford, UK.</span>,
-            date: 'August 22',
-            source: { name: 'eFestivals', title: 'Oasis really disappoint at Staffordshire\'s V Festival', url: 'https://www.efestivals.co.uk/festivals/v/stafford/2009/review-oasis.shtml' },
-            title: 'Oasis Play Last Gig',
-            type: 'gigs',
-            year: '2009'
-        },
-        {
-            description: <span>Oasis cancel an appearance at V Festival in Chelmsford, Essex, UK, claiming that singer Liam Gallagher has laryngitis.</span>,
-            date: 'August 23',
-            source: { name: 'The Guardian', title: 'Oasis pull out of V festival, as Snow Patrol step in', url: 'https://www.theguardian.com/music/2009/aug/24/oasis-snow-patrol-v-festival' },
-            title: 'Oasis Cancel V Festival Appearance',
-            type: 'gigs',
-            year: '2009'
-        },
-        {
-            description: <span>After an argument backstage with brother Liam prior to playing the Rock en Seine festival in Paris, France, Noel Gallagher quits Oasis.</span>,
-            date: 'August 28',
-            source: { name: 'BBC World News', title: 'Oasis Split as Noel Quits Group', url: 'http://news.bbc.co.uk/2/hi/entertainment/8228053.stm' },
-            title: 'Noel Gallagher Quits Oasis',
-            type: 'noteworthy',
-            year: '2009'
-        },
-        {
-            description: <span>Oasis release a compilation album of their singles titled <i>Time Flies... 1994-2009</i>, with <i>Sunday Morning Call</i> appearing as a hidden track.  It also includes a DVD disc with commentary on the band's music videos from Noel and Liam Gallagher.  The album is released on Oasis' label Big Brother.</span>,
-            date: 'June 14',
-            source: { name: 'Wikipedia', title: 'Time Flies... 1994–2009', url: 'https://en.wikipedia.org/wiki/Time_Flies..._1994%E2%80%932009' },
-            title: 'Time Flies... 1994-2009 Released',
-            type: 'releases',
-            year: '2010'
-        },
-        {
-            description: <span>Titled <i>Oasis: The Truth: My Life as Oasis's Drummer </i>, Tony McCarroll releases a book chronicling his time in Oasis as their drummer.</span>,
-            date: 'October 4',
-            source: { name: 'John Blake Books', title: 'Oasis: The Truth', url: 'https://johnblakebooks.com/oasis-the-truth.html' },
-            title: 'Tony McCarroll Releases Book About His Time in Oasis',
-            type: 'noteworthy',
-            year: '2010'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Time Flies... 1994-2009</i> album sales as platinum.</span>,
-            date: 'December 17',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Time Flies... 1994-2009 Sales Certified as Platinum by BPI',
-            type: 'certifications',
-            year: '2010'
-        },
-        {
-            description: <span>An exhibition of photographs and memorabilia titled <i>Chasing the Sun: Oasis 1993-1997</i> opens in Shoreditch, London, UK.</span>,
-            date: 'April 11',
-            source: { name: 'Q Magazine', title: 'Oasis: Chasing the Sun 1993-1997', url: 'https://www.qthemusic.com/articles/column/gallery-oasis-chasing-sun-1993-1997' },
-            title: 'Chasing the Sun Exhibition Opens in London, UK',
-            type: 'noteworthy',
-            year: '2014'
-        },
-        {
-            description: <span>As part of the <i>Chasing the Sun</i> series, a three disc version of <i>Definitely Maybe</i> is re-released to celebrate the album's 20th anniversary.  The package includes a remastered version of the album with two additional discs of b-sides and rare tracks.</span>,
-            date: 'May 19',
-            source: { name: 'Wikipedia', title: 'Definitely Maybe', url: 'https://en.wikipedia.org/wiki/Definitely_Maybe#2014_reissue' },
-            title: 'Definitely Maybe Re-Released',
-            type: 'releases',
-            year: '2014'
-        },
-        {
-            description: <span>As part of the <i>Chasing the Sun</i> series, a three disc version of <i>(What's The Story) Morning Glory?</i> is re-released to celebrate the album's upcoming 20th anniversary.  The package includes a remastered version of the album with two additional discs of b-sides and rare tracks.</span>,
-            date: 'September 29',
-            source: { name: 'Wikipedia', title: '(What\'s the Story) Morning Glory?', url: 'https://en.wikipedia.org/wiki/(What%27s_the_Story)_Morning_Glory%3F#2014_reissue' },
-            title: '(What\'s The Story) Morning Glory? Re-Released',
-            type: 'releases',
-            year: '2014'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Time Flies... 1994-2009</i> album sales as 2x platinum.</span>,
-            date: 'July 31',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Time Flies... 1994-2009 Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '2015'
-        },
-        {
-            description: <span>As part of the <i>Chasing the Sun</i> series, a three disc version of <i>Be Here Now</i> is re-released to celebrate the album's upcoming 20th anniversary.  The package includes a remastered version of the album with two additional discs of b-sides and demos and the long sought after <i>Mustique Sessions</i>.</span>,
-            date: 'October 14',
-            source: { name: 'Wikipedia', title: 'Be Here Now (album)', url: 'https://en.wikipedia.org/wiki/Be_Here_Now_(album)#2016_reissue' },
-            title: 'Be Here Now Re-Released',
-            type: 'releases',
-            year: '2016'
-        },
-        {
-            description: <span><i>Oasis: Supersonic</i> opens in theaters.  The documentary focuses on Oasis' rapid rise to fame during its first three years after signing to Creation Records.  It is directed by Mat Whitecross.</span>,
-            date: 'October 26',
-            source: { name: 'Wikipedia', title: 'Oasis: Supersonic', url: 'https://en.wikipedia.org/wiki/Oasis:_Supersonic' },
-            title: 'Oasis: Supersonic Released',
-            type: 'releases',
-            year: '2016'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Oasis: Supersonic</i> video sales as gold.</span>,
-            date: 'November 4',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified ', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Oasis: Supersonic Sales Certified as Gold by BPI',
-            type: 'certifications',
-            year: '2016'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Oasis: Supersonic</i> video sales as platinum.</span>,
-            date: 'November 11',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Oasis: Supersonic Sales Certified as Platinum by BPI',
-            type: 'certifications',
-            year: '2016'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Oasis: Supersonic</i> video sales as 2x platinum.</span>,
-            date: 'December 9',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Oasis: Supersonic Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '2016'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Oasis: Supersonic</i> video sales as 3x platinum.</span>,
-            date: 'December 30',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Oasis: Supersonic Sales Certified as 2x Platinum by BPI',
-            type: 'certifications',
-            year: '2016'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Oasis: Supersonic</i> video sales as 4x platinum.</span>,
-            date: 'October 6',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Oasis: Supersonic Sales Certified as 4x Platinum by BPI',
-            type: 'certifications',
-            year: '2017'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Time Flies... 1994-2009</i> album sales as 3x platinum.</span>,
-            date: 'March 3',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Time Flies... 1994-2009 Sales Certified as 3x Platinum by BPI',
-            type: 'certifications',
-            year: '2018'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>Time Flies... 1994-2009</i> album sales as 3x platinum.</span>,
-            date: 'March 16',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: 'Time Flies... 1994-2009 Sales Certified as 3x Platinum by BPI',
-            type: 'certifications',
-            year: '2018'
-        },
-        {
-            description: <span>The British Phonographic Institute certifies <i>(What's The Story) Morning Glory?</i> album sales as 15x platinum.</span>,
-            date: 'February 8',
-            source: { name: 'British Phonographic Industry', title: 'BRIT Certified', url: 'https://www.bpi.co.uk/brit-certified/' },
-            title: '(What\'s The Story) Morning Glory? Sales Certified as 15x Platinum by BPI',
-            type: 'certifications',
-            year: '2019'
-        },
-    ],
-
-    getFirstYear: function() {
-        return parseInt(this.data[0].year, 10);
-    },
-
-    getLastYear: function() {
-        return parseInt(this.data[this.data.length - 1].year, 10);
-    },
-
-    getNumberOfEvents: function() {
-        return this.data.length;
-    },
-
-    getNumberOfYears: function() {
-        var firstYear = this.getFirstYear();
-        var lastYear = this.getLastYear();
-        return lastYear - firstYear;
-    },
-
-    generateHistory: function(today) {
-        return this.data.filter(function(event) {
-            return event.date === today;
-        });
-    },
-
-    getIcon: function(type) {
-        switch(type) {
-          case 'certifications':
-              return 'star';
-          case 'gigs':
-              return 'music_note';
-          case 'noteworthy':
-             return 'announcement';
-          case 'photo':
-             return 'camera_alt';
-          case 'recordings':
-             return 'settings_voice';
-          case 'releases':
-             return 'album';
-          case 'videos':
-          default:
-              return 'videocam';
-        }
-    },
-
-    hasEventsForYear: function(year) {
-        return this.data.filter(e => parseInt(e.year) === parseInt(year)).length > 0;
-    }
-};
