+++ src/shared/SourceList.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import SourceUtils from '../util/sourceUtils.js';
-import TimelineData from '../data/timelineData.js';
+import TimelineData from '../data/timelineDataLoader.js';
 
 export default class SourceList extends React.Component<any, any> {
