+++ src/timeline/TimelineNavigation.tsx
@@ -5,7 +5,7 @@ import MediaQuery from "react-responsive";
 import Select from "react-select";
 import jump from 'jump.js';
 import inView from 'in-view';
-import TimelineData from '../data/timelineData.js';
+import TimelineData from '../data/timelineDataLoader.js';
 import YearScroll from '../shared/YearScroll';
 
 export default class TimelineNavigation extends React.Component<any, any> {
